Removing fixed calendar
