+++ css/styles.css
@@ -404,6 +404,11 @@ body {
   color: var(--punch-disable-bground);
 }
 
+#punch-button-label {
+  cursor: pointer;
+  margin: auto;
+}
+
 .waiver-img {
   display: none;
 }
@@ -725,163 +730,7 @@ input:disabled + .slider {
   background-position: center;
 }
 
-/** Day view related styles **/
-
-html[data-view='day'] .container {
-  margin-bottom: 0px;
-}
-
-html[data-view='day'] .title-header {
-  width: 100%;
-  margin-bottom: 80px;
-}
-
-html[data-view='day'] .title-header-text {
-  font-size: 300%;
-}
-
-html[data-view='day'] .table-header {
-  width: 100%;
-}
-
-html[data-view='day'] #calendar input[type="date" i]::-webkit-calendar-picker-indicator {
-  background-image: var(--day-calendar-select-icon);
-  height: 10px;
-  margin-left: 4px;
-  width: 10px;
-}
-
-html[data-view='day'] #calendar-table-body {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  margin-top: 10px;
-}
-
-html[data-view='day'] #calendar-table-body > div {
-  width: 100%;
-  display: flex;
-  flex: auto;
-  flex-direction: row;
-}
-
-html[data-view='day'] .table-header tr {
-  border-bottom: 2px solid var(--punch-bground);
-}
-
-html[data-view='day'] .th-label {
-  text-align: right;
-  padding-right: 20px;
-  width: 50%;
-}
-
-html[data-view='day'] .th-label.th-label-total {
-  width: 33%;
-}
-
-html[data-view='day'] .but-switch-view {
-  width: 6.5%;
-}
-
-html[data-view='day'] .ti {
-  text-align: left;
-  padding-left: 20px;
-  width: 50%;
-}
-
-html[data-view='day'] .ti-total {
-  width: 13%;
-  overflow: hidden;
-}
-
-html[data-view='day'] .ti-total input {
-  margin-left: -14px;
-}
-
-html[data-view='day'] .table-body {
-  margin-top: 10px;
-}
-
-html[data-view='day'] .row-time {
-  border-bottom: 1px solid var(--table-border);
-  width: 90% !important;
-  margin: auto;
-  padding: 2px 0 2px 0;
-}
-
-html[data-view='day'] .row-total {
-  border-bottom: 4px solid var(--table-total-border);
-  padding: 10px 0 10px 0;
-}
-
-html[data-view='day'] .row-waiver {
-  margin: auto;
-  padding: 2px 0 12px 0;
-  border-bottom: 4px solid var(--table-total-border);
-}
-
-html[data-view='day'] .row-waiver .waived-day-text {
-  margin: auto 20px auto auto;
-}
-
-html[data-view='day'] .row-waiver .ti-total {
-  margin-right: auto;
-}
-
-html[data-view='day'] .non-working-day {
-  background: none;
-  margin: auto auto 10px auto;
-}
-
-html[data-view='day'] .summary {
-  padding: 5px 0 5px 0;
-}
-
-html[data-view='day'] .summary .leave-by-text {
-  margin: 0 20px auto auto;
-}
-
-html[data-view='day'] .summary .leave-by-time {
-  margin-right: auto;
-}
-
-html[data-view='day'] .month-total-row {
-  border-top: none;
-  border-bottom: none;
-  padding-top: 5px;
-}
-
-html[data-view='day'] .month-total-row .month-total-text {
-  padding-right: 10px;
-  text-align: right;
-}
-
-html[data-view='day'] .month-total-row .month-total-element {
-  margin-left: auto;
-  margin-right: auto;
-  display: flex;
-  width: fit-content;
-}
-
-html[data-view='day'] .month-total-row .month-total-time {
-  text-align: left;
-}
-
-html[data-view='day'] .footer {
-  height: 32px;
-  font-size: 100%;
-}
-
-html[data-view='day'] .footer .punch-button {
-  font-weight: 700;
-  font-family: 'Montserrat', sans-serif;
-}
-
-html[data-view='day'] .punch-button img {
-  display: none;
-}
-
-/** Extended view related styles **/
+/** Flexible view related styles **/
 
 html[data-view='flexible'] .calendar-table-header {
   display: flow-root;
