+++ js/saved-preferences.js
@@ -1,26 +1,13 @@
 'use strict';
 const i18n = require('../src/configs/i18next.config');
 
-
 const { app, ipcMain } = require('electron');
 
 let savedPreferences = null;
-let alreadyAskedForFlexibleDbMigration = false;
 
 function getSavedPreferences()
 {
     return savedPreferences;
-}
-
-// Sets if TTL already asked for migration. True means it did.
-function setAlreadyAskedForFlexibleDbMigration(opt)
-{
-    alreadyAskedForFlexibleDbMigration = opt;
-}
-
-function getAlreadyAskedForFlexibleDbMigration()
-{
-    return alreadyAskedForFlexibleDbMigration;
 }
 
 ipcMain.on('PREFERENCE_SAVE_DATA_NEEDED', (event, preferences) =>
@@ -36,7 +23,5 @@ ipcMain.on('PREFERENCE_SAVE_DATA_NEEDED', (event, preferences) =>
 });
 
 module.exports = {
-    getSavedPreferences,
-    setAlreadyAskedForFlexibleDbMigration,
-    getAlreadyAskedForFlexibleDbMigration
+    getSavedPreferences
 };
