+++ locales/ca-CA/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "Buidar la base de dades",
         "confirm-clear-all-data": "¿Segur que vols esborrar totes les dades?",
         "confirm-import-db": "¿Segur que vols importar una base de dates? Anularás qualsevol informació actual.",
-        "confirm-migrate-to-flexible": "Segur que vols migrar la base de dades del calendari fix a el calendari flexible?\n\nLa base de dades del calendari flexible existent se suprimirà.",
         "copy": "Copiat",
         "could-not-be-loaded": "no ha pogut ser carregat",
         "cut": "Tallar",
@@ -59,14 +58,16 @@
         "export": "Exportar",
         "failed-entries": "Entrades fallides",
         "failed-migrating": "Migració fallida",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "Ajuda",
         "import-database": "Importar base de dades",
         "import-db-from-file": "Importar base de dades d'un arxiu",
         "import-successful": "¡Importació exitosa!",
         "import": "Importar",
         "menu": "Menu",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "Migrar base de dades del calendari fixe a calendari flexible",
-        "migrate-to-flexible-calendar": "Migrar a calendari flexible",
         "migration-successful": "¡Migració exitosa!",
         "no-thanks": "No, gracies",
         "ok": "OK",
@@ -76,9 +77,10 @@
         "punched-time": "Temps fitxat",
         "quit": "Sortir",
         "reload": "Recarregar",
+        "report": "Report",
         "select-all": "Seleccionar tot",
         "send-feedback": "Envieu els vostres comentaris",
-        "should-migrate-to-flexible": "El teu calendari flexible està buit. Vols començar per migrar la base de dades del calendari fix existent a la flexible?",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "Mostra Aplicació",
         "something-went-wrong": "Hi ha hagut un error :(",
         "toggle-dev-tools": "Alternar a eines de desenvolupament",
@@ -88,17 +90,10 @@
         "yes-please": "Sí, si us plau"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "Canviar la vista",
         "previous-day": "Dia anterior",
         "next-day": "Dia següent",
-        "time-to-leave": "Time to Leave",
         "current-day": "Anar al dia actual",
-        "day-done-balance": "Todo a punt per avui. El balanç del día:",
         "leave-by": "Hauries de sortir al les:",
-        "month-balance-title": "Balanç a dia d'avui per aquest mes. Un balanç positiu significa hores extres que no haurás de treballar avui (o el que resta de mes).",
-        "month-balance": "Balanç del mes",
-        "overall-balance-title": "Balanç global fins fi de mes o del dia actual",
-        "overall-balance": "Balanç general",
         "not-a-working-day": "No es un día de trabajo",
         "waived-day": "Dia de descans",
         "day-total": "Total del dia",
@@ -109,23 +104,16 @@
         "entry": "Entrada"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "Canviar la vista",
         "current-month": "Anar al mes actual",
         "previous-month": "Mes anterior",
         "next-month": "Mes següent",
-        "time-to-leave": "Time to Leave",
         "day": "Dia",
         "total": "Total",
         "leave-by": "Basant-se en l'hora en què va arribar avui, hauria de marxar per",
-        "day-done-balance": "Tot a punt per avui. El Balanç del dia:",
         "last-day-balance": "Últim dia utilitzat per al Balanç",
         "working-days-title": "Quants dies feiners hi ha al mes",
-        "month-balance-title": "Balanç fins avui per a aquest mes. Un Balanç positiu significa hores extres que no necessites treballar avui (o la resta del mes).",
-        "overall-balance-title": "Balanç general fins al final del mes o d'avui",
         "on": "Pres",
         "working-days": "Dies de treball",
-        "month-balance": "Balanç del mes",
-        "overall-balance": "Balanç general",
         "add-waiver-day": "Afegeix dia lliure",
         "waived-day": "Dia lliure",
         "remove-entry": "Eliminar l'entrada",
@@ -134,52 +122,17 @@
         "no": "No"
     },
     "$Calendar": {
-        "leave-by": "Basant-se en l'hora en què va arribar avui, hauria de marxar per",
-        "all-done": "Tot a punt per avui. El Balanç del dia:",
-        "last-day-balance": "Últim dia utilitzat per al Balanç",
-        "on": "Pres",
-        "working-days": "Dies de treball",
-        "working-days-title": "Quants dies feiners hi ha al mes",
+        "day-done-balance": "Todo a punt per avui. El balanç del día:",
         "month-balance": "Balanç del mes",
         "month-balance-title": "Balanç fins avui per a aquest mes. Un Balanç positiu significa hores extres que no necessites treballar avui (o la resta del mes).",
         "overall-balance": "Balanç general",
         "overall-balance-title": "Balanç global fins al final del mes o del dia actual",
         "switch-view": "Canviar la vista",
-        "switch-view-title": "Canviar la vista",
-        "current-month": "Mes actual",
-        "current-month-title": "Anar a al mes actual",
-        "previous-month": "Mes anterior",
-        "next-month": "Mes següent",
-        "time-to-leave": "Time to Leave",
-        "day-start": "Inici de el dia",
-        "lunch-begin": "Inici de dinar",
-        "lunch-end": "Cap de l'esmorzar",
-        "day-end": "Fi del dia",
-        "lunch-total": "Total dinar",
-        "day-total": "Total del dia",
-        "day": "Dia",
-        "lunch-start": "Començament de l'esmorzar",
-        "punch-time": "Hora de Fitxar"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "¡Hola! Crec que es hora de marcha",
         "punch-reminder": "No T'oblidis de fitxar!"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "Hauries de anar-te'n:",
-        "all-done": "Tot a punt per avui. El Balanç del dia:",
-        "month-balance": "Balanç del mes",
-        "month-balance-title": "Balanç fins avui per a aquest mes. Un Balanç positiu significa hores extres que no necessites treballar avui (o la resta del mes).",
-        "overall-balance": "Balanç general",
-        "overall-balance-title": "Balanç global fins al final del mes o del dia actual",
-        "non-working-day": "No és un dia de treball",
-        "waived-day": "Dia de lliure",
-        "day-start": "Inici de el dia",
-        "lunch-begin": "Inici de dinaroverallBalanceStart",
-        "lunch-end": "Cap de l'esmorzar",
-        "day-end": "Fi del dia",
-        "lunch-total": "Total dinar",
-        "day-total": "Total del dia"
     },
     "$WorkdayWaiver": {
         "add-waiver": "Afegeix dia lliure",
