+++ locales/de-DE/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "Datenbank leeren",
         "confirm-clear-all-data": "Sind Sie sich sicher, dass Sie alle Daten löschen möchten?",
         "confirm-import-db": "Sind Sie sich sicher, dass Sie eine Datenbank importieren möchten? Alle vorhandenen Daten werden überschrieben.",
-        "confirm-migrate-to-flexible": "Sind Sie sich sicher, dass Sie die fixierte Kalender Datenbank in eine flexible Kalender Datenbank migrieren möchten?\n\nDie existierende flexible Kalender Datenbank wird geleert.",
         "copy": "Kopieren",
         "could-not-be-loaded": "konnte nicht geladen werden",
         "cut": "Ausschneiden",
@@ -59,14 +58,16 @@
         "export": "Export",
         "failed-entries": "Fehlerhafte Einträge",
         "failed-migrating": "Migrierung fehlgeschlagen",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "Hilfe",
         "import-database": "Datenbank Import",
         "import-db-from-file": "Datenbank Import aus Datei",
         "import-successful": "Yay! Import erfolgreich!",
         "import": "Import",
         "menu": "Menü",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "Migriere fixierte Kalender Datenbank zu flexibler",
-        "migrate-to-flexible-calendar": "Migriere zu flexiblen Kalender",
         "migration-successful": "Yay! Migration erfolgreich!",
         "no-thanks": "Nein, danke",
         "ok": "OK",
@@ -76,9 +77,10 @@
         "punched-time": "Gestoppte Zeit",
         "quit": "Schließen",
         "reload": "Aktualisieren",
+        "report": "Melden",
         "select-all": "Alle selektieren",
         "send-feedback": "Feedback senden",
-        "should-migrate-to-flexible": "Ihr flexibler Kalender ist leer. Möchten Sie die existierende fixierte Kalender Datenbank in eine flexible migrieren?",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "Zeige App",
         "something-went-wrong": "Etwas ist falsch gelaufen :(",
         "toggle-dev-tools": "Entwickler Werkzeuge umschalten",
@@ -88,17 +90,10 @@
         "yes-please": "Ja, bitte"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "Ansicht ändern",
         "current-day": "Zum heutigen Tag springen",
         "previous-day": "Vorheriger Tag",
         "next-day": "Nächster Tag",
-        "time-to-leave": "Time to Leave",
         "leave-by": "Sie sollten gehen um:",
-        "day-done-balance": "Alles erledigt für heute. Tagesbilanz:",
-        "month-balance": "Monatsbilanz",
-        "month-balance-title": "Bilanz positiv bis heute für diesen Monat. Eine positive Bilanz bedeutet extra Stunden für die Sie heute nicht arbeiten müssen (oder den Rest des Monats).",
-        "overall-balance": "Gesamtbilanz",
-        "overall-balance-title": "Gesamtbilanz bis zum Ende des Monats oder den heutigen Tag",
         "not-a-working-day": "Kein Arbeitstag",
         "waived-day": "Ausgeschlossener Tag",
         "day-total": "Tag gesamt",
@@ -109,23 +104,16 @@
         "entry": "Eintrag"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "Ansicht ändern",
         "current-month": "Zum heutigen Monat springen",
         "previous-month": "Vorheriger Monat",
         "next-month": "Nächster Monat",
-        "time-to-leave": "Time to Leave",
         "day": "Tag",
         "total": "Gesamt",
         "leave-by": "Basierend auf der Zeit an der Sie heute angekommen sind, sollten Sie zur folgenden Uhrzeit gehen",
-        "day-done-balance": "Alles erledigt für heute. Tagesbilanz:",
         "last-day-balance": "Letzten Tag genutzt für Bilanz",
         "working-days-title": "Anzahl der Arbeitstage in diesem Monat",
-        "month-balance-title": "Bilanz positiv bis heute für diesen Monat. Eine positive Bilanz bedeutet extra Stunden für die Sie heute nicht arbeiten müssen (oder den Rest des Monats).",
-        "overall-balance-title": "Gesamtbilanz bis zum Ende des Monats oder den heutigen Tag",
         "on": "An",
         "working-days": "Arbeitstage",
-        "month-balance": "Monatsbilanz",
-        "overall-balance": "Gesamtbilanz",
         "add-waiver-day": "Füge einen Ausschluss für diesen Tag hinzu",
         "waived-day": "Ausgeschlossener Tag",
         "remove-entry": "Eintrag löschen",
@@ -134,52 +122,17 @@
         "no": "Nein"
     },
     "$Calendar": {
-        "leave-by": "Basierend auf der Zeit an der Sie heute angekommen sind, sollten Sie zur folgenden Uhrzeit gehen",
-        "all-done": "Alles erledigt für heute. Tagesbilanz:",
-        "last-day-balance": "Letzten Tag genutzt für Bilanz",
-        "on": "An",
-        "working-days": "Arbeitstage",
-        "working-days-title": "Anzahl der Arbeitstage in diesem Monat",
+        "day-done-balance": "Alles erledigt für heute. Tagesbilanz:",
         "month-balance": "Monatsbilanz",
         "month-balance-title": "Bilanz positiv bis heute für diesen Monat. Eine positive Bilanz bedeutet extra Stunden für die Sie heute nicht arbeiten müssen (oder den Rest des Monats).",
         "overall-balance": "Gesamtbilanz",
         "overall-balance-title": "Gesamtbilanz bis zum Ende des Monats oder den heutigen Tag",
         "switch-view": "Ansicht ändern",
-        "switch-view-title": "Ansicht ändern",
-        "current-month": "Aktueller Monat",
-        "current-month-title": "Zum heutigen Monat springen",
-        "previous-month": "Vorheriger Monat",
-        "next-month": "Nächster Monat",
-        "time-to-leave": "Time to Leave",
-        "day-start": "Tag Beginn",
-        "lunch-begin": "Mittag Beginn",
-        "lunch-end": "Mittag Ende",
-        "day-end": "Tag Ende",
-        "lunch-total": "Mittag Gesamt",
-        "day-total": "Tag Gesamt",
-        "day": "Tag",
-        "lunch-start": "Mittag Beginn",
-        "punch-time": "Punch Time"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "Sie da! Ich denke es ist Zeit zu gehen.",
         "punch-reminder": "Nicht vergessen Zeiten einzutragen"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "Sie sollten gehen um:",
-        "all-done": "Alles erledigt für heute. Tagesbilanz:",
-        "month-balance": "Monatsbilanz",
-        "month-balance-title": "Bilanz positiv bis heute für diesen Monat. Eine positive Bilanz bedeutet extra Stunden für die Sie heute nicht arbeiten müssen (oder den Rest des Monats).",
-        "overall-balance": "Gesamtbilanz",
-        "overall-balance-title": "Gesamtbilanz bis zum Ende des Monats oder den heutigen Tag",
-        "non-working-day": "Kein Arbeitstag",
-        "waived-day": "Ausgeschlossener Tag",
-        "day-start": "Tag Beginn",
-        "lunch-begin": "Mittag Beginn",
-        "lunch-end": "Mittag Ende",
-        "day-end": "Tag Ende",
-        "lunch-total": "Mittag Gesamt",
-        "day-total": "Tag Gesamt"
     },
     "$WorkdayWaiver": {
         "add-waiver": "Ausschluss hinzufügen",
