+++ locales/en/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "Clear Database",
         "confirm-clear-all-data": "Are you sure you want to clear all the data?",
         "confirm-import-db": "Are you sure you want to import a database? It will override any current information.",
-        "confirm-migrate-to-flexible": "Are you sure you want to migrate the fixed calendar database to the flexible calendar?\n\nThe existing flexible calendar database will be cleared.",
         "copy": "Copy",
         "could-not-be-loaded": "could not be loaded",
         "cut": "Cut",
@@ -59,14 +58,16 @@
         "export": "Export",
         "failed-entries": "Failed entries",
         "failed-migrating": "Failed migrating",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "Help",
         "import-database": "Import database",
         "import-db-from-file": "Import database from file",
         "import-successful": "Yay! Import successful!",
         "import": "Import",
         "menu": "Menu",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "Migrate fixed calendar database to flexible",
-        "migrate-to-flexible-calendar": "Migrate to flexible calendar",
         "migration-successful": "Yay! Migration successful!",
         "no-thanks": "No, thanks",
         "ok": "OK",
@@ -76,9 +77,10 @@
         "punched-time": "Punched time",
         "quit": "Quit",
         "reload": "Reload",
+        "report": "Report",
         "select-all": "Select All",
         "send-feedback": "Send feedback",
-        "should-migrate-to-flexible": "Your flexible calendar is empty. Do you want to start by migrating the existing fixed calendar database to your flexible one?",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "Show App",
         "something-went-wrong": "Something wrong happened :(",
         "toggle-dev-tools": "Toggle Developer Tools",
@@ -88,17 +90,10 @@
         "yes-please": "Yes, please"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "Switch View",
         "current-day": "Go to Current Day",
         "previous-day": "Previous Day",
         "next-day": "Next Day",
-        "time-to-leave": "Time to Leave",
         "leave-by": "You should leave by:",
-        "day-done-balance": "All done for today. Balance of the day:",
-        "month-balance": "Month Balance",
-        "month-balance-title": "Balance up until today for this month. A positive balance means extra hours you don't need to work today (or the rest of the month).",
-        "overall-balance": "Overall Balance",
-        "overall-balance-title": "Overall balance until end of the month or current day",
         "not-a-working-day": "Not a working day",
         "waived-day": "Waived day",
         "day-total": "Day Total",
@@ -109,23 +104,16 @@
         "entry": "Entry"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "Switch View",
         "current-month": "Go to Current Month",
         "previous-month": "Previous Month",
         "next-month": "Next Month",
-        "time-to-leave": "Time to Leave",
         "day": "Day",
         "total": "Total",
         "leave-by": "Based on the time you arrived today, you should leave by",
-        "day-done-balance": "All done for today. Balance of the day:",
         "last-day-balance": "Last day used for balance",
         "on": "On",
         "working-days": "Working days",
         "working-days-title": "How many working days there's in the month",
-        "month-balance": "Month Balance",
-        "month-balance-title": "Balance up until today for this month. A positive balance means extra hours you don't need to work today (or the rest of the month).",
-        "overall-balance": "Overall Balance",
-        "overall-balance-title": "Overall balance until end of the month or current day",
         "add-waiver-day": "Add a waiver for this day",
         "waived-day": "Waived day",
         "remove-entry": "Remove entry",
@@ -133,53 +121,18 @@
         "yes": "Yes",
         "no": "No"
     },
-    "$Calendar": {
-        "leave-by": "Based on the time you arrived today, you should leave by",
-        "all-done": "All done for today. Balance of the day:",
-        "last-day-balance": "Last day used for balance",
-        "on": "On",
-        "working-days": "Working days",
-        "working-days-title": "How many working days there's in the month",
+    "$BaseCalendar": {
+        "day-done-balance": "All done for today. Balance of the day:",
         "month-balance": "Month Balance",
         "month-balance-title": "Balance up until today for this month. A positive balance means extra hours you don't need to work today (or the rest of the month).",
         "overall-balance": "Overall Balance",
         "overall-balance-title": "Overall balance until end of the month or current day",
         "switch-view": "Switch View",
-        "switch-view-title": "Switch View",
-        "current-month": "Current Month",
-        "current-month-title": "Go to Current Month",
-        "previous-month": "Previous Month",
-        "next-month": "Next Month",
-        "time-to-leave": "Time to Leave",
-        "day-start": "Day Start",
-        "lunch-begin": "Lunch Begin",
-        "lunch-end": "Lunch End",
-        "day-end": "Day End",
-        "lunch-total": "Lunch Total",
-        "day-total": "Day Total",
-        "day": "Day",
-        "lunch-start": "Lunch Start",
-        "punch-time": "Punch time"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "Hey there! I think it's time to leave.",
         "punch-reminder": "Don't forget to punch in!"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "You should leave by:",
-        "all-done": "All done for today. Balance of the day:",
-        "month-balance": "Month Balance",
-        "month-balance-title": "Balance up until today for this month. A positive balance means extra hours you don't need to work today (or the rest of the month).",
-        "overall-balance": "Overall balance",
-        "overall-balance-title": "Overall balance until end of the month or current day",
-        "non-working-day": "Not a working day",
-        "waived-day": "Waived day",
-        "day-start": "Day Start",
-        "lunch-begin": "Lunch Begin",
-        "lunch-end": "Lunch End",
-        "day-end": "Day End",
-        "lunch-total": "Lunch Total",
-        "day-total": "Day Total"
     },
     "$WorkdayWaiver": {
         "add-waiver": "Add Waiver",
