+++ locales/es/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "Borrar la base de datos",
         "confirm-clear-all-data": "¿Estás seguro de que quieres borrar todos los datos?",
         "confirm-import-db": "¿Estás seguro de que quieres importar una base de datos? Anularás cualquier información actual.",
-        "confirm-migrate-to-flexible": "¿Estás seguro de que quieres migrar la base de datos del calendario fijo al calendario flexible?\n\nLa base de datos del calendario flexible existente será eliminada.",
         "copy": "Copiado",
         "could-not-be-loaded": "no pudo ser cargado",
         "cut": "Cortar",
@@ -59,14 +58,16 @@
         "export": "Exportar",
         "failed-entries": "Entradas fallidas",
         "failed-migrating": "Migración fallida",
+        "fresh-start": "Empezar desde cero",
+        "fresh-start-confirm": "¿Estás seguro de que quieres empezar desde cero?",
         "help": "Ayuda",
         "import-database": "Importar base de datos",
         "import-db-from-file": "Importar base de datos de un archivo",
         "import-successful": "¡Sí! ¡Importación exitosa!",
         "import": "Importar",
         "menu": "Menú",
+        "migrate": "Migrar",
         "migrate-calendar-to-flexible": "Migrar base de datos de calendario fijo a calendario flexible",
-        "migrate-to-flexible-calendar": "Migrar a calendario flexible",
         "migration-successful": "Sí! ¡Migración exitosa!",
         "no-thanks": "No, gracias",
         "ok": "OK",
@@ -76,9 +77,10 @@
         "punched-time": "Tiempo fichado",
         "quit": "Salir",
         "reload": "Recargar",
+        "report": "Reportar",
         "select-all": "Seleccionar Todo",
         "send-feedback": "Envíe sus comentarios",
-        "should-migrate-to-flexible": "Tu calendario flexible está vacío. ¿Quieres empezar por migrar la base de datos del calendario fijo existente a la flexible?",
+        "should-migrate-to-flexible": "Una de las novedades de este lanzamiento de Time to Leave es un número flexible de entradas para el calendario.\n\nEsta nueva vista usa un nuevo formato en la base de datos.\nPara usarlo, puedes migrar su base de datos existente al nuevo formato, o empezar desde cero.",
         "show-app": "Mostrar Aplicacion",
         "something-went-wrong": "Algo malo pasó :(",
         "toggle-dev-tools": "Alternar a herramientas de desarrollo",
@@ -88,17 +90,11 @@
         "yes-please": "Sí, por favor"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "Cambiar la vista",
         "previous-day": "Día anterior",
         "next-day": "Día siguiente",
-        "time-to-leave": "Time to Leave",
         "current-day": "Ir a dia actual",
-        "day-done-balance": "Todo listo por hoy. El balance del día:",
         "leave-by": "Deberías irte a las:",
-        "month-balance-title": "Balance hasta hoy para este mes. Un balance positivo significa horas extras que no necesitas trabajar hoy (o el resto del mes).",
         "month-balance": "Balance del mes",
-        "overall-balance-title": "Balance global hasta el final del mes o del día actual",
-        "overall-balance": "Balance general",
         "not-a-working-day": "No es un día de trabajo",
         "waived-day": "Día de descanso",
         "day-total": "Total del día",
@@ -109,23 +105,17 @@
         "entry": "Entrada"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "Cambiar la vista",
         "current-month": "Ir al mes actual",
         "previous-month": "Mes anterior",
         "next-month": "Mes siguiente",
-        "time-to-leave": "Time to Leave",
         "day": "Día",
         "total": "Total",
         "leave-by": "Basándose en la hora en que llegó hoy, debería irse por",
-        "day-done-balance": "Todo listo por hoy. El balance del día:",
         "last-day-balance": "Último día utilizado para el balance",
         "working-days-title": "Cuántos días laborables hay en el mes",
-        "month-balance-title": "Balance hasta hoy para este mes. Un balance positivo significa horas extras que no necesitas trabajar hoy (o el resto del mes).",
-        "overall-balance-title": "Balance general hasta el final del mes o de hoy",
         "on": "Prendido",
         "working-days": "Días de trabajo",
         "month-balance": "Balance del mes",
-        "overall-balance": "Balance general",
         "add-waiver-day": "Añade día libre",
         "waived-day": "Día libre",
         "remove-entry": "Eliminar la entrada",
@@ -133,53 +123,18 @@
         "yes": "Sí",
         "no": "No"
     },
-    "$Calendar": {
-        "leave-by": "Basándose en la hora en que llegó hoy, debería irse por",
-        "all-done": "Todo listo por hoy. El balance del día:",
-        "last-day-balance": "Último día utilizado para el balance",
-        "on": "Prendido",
-        "working-days": "Días de trabajo",
-        "working-days-title": "Cuántos días laborables hay en el mes",
+    "$BaseCalendar": {
+        "day-done-balance": "Todo listo por hoy. El balance del día:",
         "month-balance": "Balance del mes",
         "month-balance-title": "Balance hasta hoy para este mes. Un balance positivo significa horas extras que no necesitas trabajar hoy (o el resto del mes).",
         "overall-balance": "Balance general",
         "overall-balance-title": "Balance global hasta el final del mes o del día actual",
         "switch-view": "Cambiar la vista",
-        "switch-view-title": "Cambiar la vista",
-        "current-month": "Mes actual",
-        "current-month-title": "Ir al mes actual",
-        "previous-month": "Mes anterior",
-        "next-month": "Mes siguiente",
-        "time-to-leave": "Time to Leave",
-        "day-start": "Inicio del día",
-        "lunch-begin": "Inicio de almuerzo",
-        "lunch-end": "Fin del almuerzo",
-        "day-end": "Fin del día",
-        "lunch-total": "Total del almuerzo",
-        "day-total": "Total del día",
-        "day": "Día",
-        "lunch-start": "Comienzo del almuerzo",
-        "punch-time": "Hora del golpe"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "¡Hola! Creo que es hora de irse",
         "punch-reminder": "¡No olvides marcar tarjeta!"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "Deberías irte:",
-        "all-done": "Todo listo por hoy. El balance del día:",
-        "month-balance": "Balance del mes",
-        "month-balance-title": "Balance hasta hoy para este mes. Un balance positivo significa horas extras que no necesitas trabajar hoy (o el resto del mes).",
-        "overall-balance": "Balance general",
-        "overall-balance-title": "Balance global hasta el final del mes o del día actual",
-        "non-working-day": "No es un día de trabajo",
-        "waived-day": "Día de libre",
-        "day-start": "Inicio del día",
-        "lunch-begin": "Inicio de almuerzo",
-        "lunch-end": "Fin del almuerzo",
-        "day-end": "Fin del día",
-        "lunch-total": "Total del almuerzo",
-        "day-total": "Total del día"
     },
     "$WorkdayWaiver": {
         "add-waiver": "Añade día libre",
