+++ locales/fr-FR/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "Effacer les données",
         "confirm-clear-all-data": "Voulez-vous vraiment effacer toutes les données?",
         "confirm-import-db": "Voulez-vous vraiment importer les données? Il remplacera toutes les informations actuelles.",
-        "confirm-migrate-to-flexible": "Voulez-vous vraiment migrer les données du calendrier fixe vers le calendrier flexible?\n\nLes données du calendrier flexible existantes seront effacées.",
         "copy": "Copier",
         "could-not-be-loaded": "n'a pas pu être chargé",
         "cut": "Couper",
@@ -59,14 +58,16 @@
         "export": "Exporter",
         "failed-entries": "Entrées échouées",
         "failed-migrating": "Échec de la migration",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "Aide",
         "import-database": "Importer la base de données",
         "import-db-from-file": "Importer la base de données à partir d'un fichier",
         "import-successful": "youpi! Importation réussie!",
         "import": "Importer",
         "menu": "Menu",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "Migrer la base de données de calendrier fixe vers flexible",
-        "migrate-to-flexible-calendar": "Migrer vers un calendrier flexible",
         "migration-successful": "youpi! Migration réussie!",
         "no-thanks": "Non merci",
         "ok": "OK",
@@ -76,9 +77,10 @@
         "punched-time": "Pointage des heures",
         "quit": "Quitter",
         "reload": "Recharger",
+        "report": "Report",
         "select-all": "Tout sélectionner",
         "send-feedback": "Envoyez vos commentaires",
-        "should-migrate-to-flexible": "Votre calendrier flexible est vide. Voulez-vous commencer par migrer la base de données de calendrier fixe existante vers votre base de données flexible?",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "Afficher l'application",
         "something-went-wrong": "Quelque chose s'est mal passé :(",
         "toggle-dev-tools": "Basculer Outils de développement",
@@ -88,17 +90,10 @@
         "yes-please": "Oui s'il vous plaît"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "Changer de Vue",
         "previous-day": "Jour précédent",
         "next-day": "Jour suivant",
-        "time-to-leave": "Time to Leave",
         "current-day": "Aller au jour en cours",
-        "day-done-balance": "Tout est fait pour aujourd'hui. Balance of the day:",
         "leave-by": "Vous devriez partir d'ici:",
-        "month-balance-title": "Solde jusqu'à aujourd'hui pour ce mois. Un solde positif signifie des heures supplémentaires dont vous n'avez pas besoin de travailler pour aujourd'hui (ou le reste du mois).",
-        "month-balance": "Solde mensuel",
-        "overall-balance-title": "Solde global jusqu'à la fin du mois ou le jour en cours",
-        "overall-balance": "Solde Global",
         "not-a-working-day": "Pas un jour ouvrable",
         "waived-day": "Jour de dispense",
         "day-total": "Total jour",
@@ -109,23 +104,16 @@
         "entry": "Entrée"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "Changer de Vue",
         "current-month": "Aller au mois en cours",
         "previous-month": "Mois précédent",
         "next-month": "Mois suivant",
-        "time-to-leave": "Time to Leave",
         "day": "Jour",
         "total": "Total",
         "leave-by": "En fonction de l'heure à laquelle vous êtes arrivé aujourd'hui, vous devriez partir d'ici",
-        "day-done-balance": "Tout est fait pour aujourd'hui. Bilan journée:",
         "last-day-balance": "Dernier jour utilisé pour Solde",
         "on": "Sur",
         "working-days": "Jours de travail",
         "working-days-title": "Combien de jours ouvrables y a t-il dans le mois",
-        "month-balance": "Solde mensuel",
-        "month-balance-title": "Solde jusqu'à aujourd'hui pour ce mois. Un solde positif signifie des heures supplémentaires dont vous n'avez pas besoin de travailler pour aujourd'hui (ou le reste du mois).",
-        "overall-balance": "Solde Global",
-        "overall-balance-title": "Solde global jusqu'à la fin du mois ou le jour en cours",
         "add-waiver-day": "Ajouter une dispense pour ce jour",
         "waived-day": "Jour de dispense",
         "remove-entry": "Supprimer l'entrée",
@@ -134,52 +122,17 @@
         "no": "Non"
     },
     "$Calendar": {
-        "leave-by": "En fonction de l'heure à laquelle vous êtes arrivé aujourd'hui, vous devriez partir d'ici",
-        "all-done": "Tout est fait pour aujourd'hui. Bilan journée:",
-        "last-day-balance": "Dernier jour utilisé pour Solde",
-        "on": "Sur",
-        "working-days": "Jours de travail",
-        "working-days-title": "Combien de jours ouvrables y a t-il dans le mois",
+        "day-done-balance": "Tout est fait pour aujourd'hui. Bilan journée:",
         "month-balance": "Solde mensuel",
         "month-balance-title": "Solde jusqu'à aujourd'hui pour ce mois. Un solde positif signifie des heures supplémentaires dont vous n'avez pas besoin de travailler pour aujourd'hui (ou le reste du mois).",
         "overall-balance": "Solde Global",
         "overall-balance-title": "Solde global jusqu'à la fin du mois ou le jour en cours",
         "switch-view": "Changer de Vue",
-        "switch-view-title": "Changer de Vue",
-        "current-month": "Mois en cours",
-        "current-month-title": "Aller au mois en cours",
-        "previous-month": "Mois Précédent",
-        "next-month": "Mois Suivant",
-        "time-to-leave": "Time to Leave",
-        "day-start": "Début Journée",
-        "lunch-begin": "Début Déjeuner",
-        "lunch-end": "Fin Déjeuner",
-        "day-end": "Fin Journée",
-        "lunch-total": "Total Déjeuner",
-        "day-total": "Total jour",
-        "day": "Jour",
-        "lunch-start": "Début Déjeuner",
-        "punch-time": "Pointer l'heure"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "Héy! Je pense qu'il est temps de partir.",
         "punch-reminder": "N'oubliez pas de pointer!"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "Vous devriez partir d'ici:",
-        "all-done": "Tout est fait pour aujourd'hui. Bilan journée:",
-        "month-balance": "Solde mensuel",
-        "month-balance-title": "Solde jusqu'à aujourd'hui pour ce mois. Un solde positif signifie des heures supplémentaires dont vous n'avez pas besoin de travailler pour aujourd'hui (ou le reste du mois).",
-        "overall-balance": "Solde Global",
-        "overall-balance-title": "Solde global jusqu'à la fin du mois ou le jour en cours",
-        "non-working-day": "Pas un jour ouvrable",
-        "waived-day": "Jour de dispense",
-        "day-start": "Début Journée",
-        "lunch-begin": "Début Déjeuner",
-        "lunch-end": "Fin Déjeuner",
-        "day-end": "Fin Journée",
-        "lunch-total": "Total Déjeuner",
-        "day-total": "Total jour"
     },
     "$WorkdayWaiver": {
         "add-waiver": "Ajouter une dispense",
