+++ locales/hi/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "डेटाबेस साफ़ करें",
         "confirm-clear-all-data": "क्या आप वाकई सभी डेटा को साफ़ करना चाहते हैं?",
         "confirm-import-db": "क्या आप वाकई डेटाबेस आयात करना चाहते हैं? यह किसी भी वर्तमान जानकारी को ओवरराइड करेगा।",
-        "confirm-migrate-to-flexible": "क्या आप वाकई कैलेंडर कैलेंडर को लचीले कैलेंडर में माइग्रेट करना चाहते हैं? मौजूदा लचीले कैलेंडर डेटाबेस को साफ़ कर दिया जाएगा।",
         "copy": "प्रतिलिपि",
         "could-not-be-loaded": "लादा नहीं जा सकता था",
         "cut": "कट गया",
@@ -59,14 +58,16 @@
         "export-db-to-file": "फाइल करने के लिए डेटाबेस निर्यात करें",
         "failed-entries": "प्रविष्टियां विफल",
         "failed-migrating": "पलायन करने में विफल",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "मदद",
         "import": "आयात",
         "import-database": "डेटाबेस आयात करें",
         "import-db-from-file": "डेटाबेस फ़ाइल से आयात करें",
         "import-successful": "वाह! आयात सफल!",
         "menu": "मेन्यू",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "तय कैलेंडर डेटाबेस को लचीले में माइग्रेट करें",
-        "migrate-to-flexible-calendar": "लचीले कैलेंडर में माइग्रेट करें",
         "migration-successful": "वाह! प्रवासन सफल!",
         "no-thanks": "जी नहीं, धन्यवाद",
         "ok": "ठीक",
@@ -76,9 +77,10 @@
         "punched-time": "छिद्रित समय",
         "quit": "छोड़ना",
         "reload": "पुनः लोड करें",
+        "report": "Report",
         "select-all": "सभी का चयन करे",
         "send-feedback": "प्रतिक्रिया भेजें",
-        "should-migrate-to-flexible": "आपका लचीला कैलेंडर खाली है। क्या आप मौजूदा तय कैलेंडर डेटाबेस को अपने लचीले एक पर माइग्रेट करके शुरू करना चाहते हैं?",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "ऐप दिखाएं",
         "something-went-wrong": "कुछ ग़लत हुआ :(",
         "toggle-dev-tools": "डेवलपर टूल टॉगल करें",
@@ -89,22 +91,16 @@
     },
     "$FlexibleDayCalendar": {
         "current-day": "करंट डे पर जाएं",
-        "day-done-balance": "सब आज के लिए किया। दिन का संतुलन:",
         "day-total": "दिन का कुल",
         "entry": "प्रवेश",
         "entry-removal-confirmation": "क्या आप वाकई अंतिम प्रविष्टि पंक्ति को हटाना चाहते हैं?",
         "leave-by": "आपको इसके द्वारा छोड़ देना चाहिए:",
         "month-balance": "महीना शेष",
-        "month-balance-title": "इस महीने के लिए आज तक शेष राशि। एक सकारात्मक संतुलन का मतलब है अतिरिक्त घंटे जो आपको आज (या बाकी महीने) काम करने की आवश्यकता नहीं है।",
         "next-day": "अगले दिन",
         "no": "नहीं",
         "not-a-working-day": "काम का दिन नहीं",
-        "overall-balance": "समग्र संतुलन",
-        "overall-balance-title": "महीने के अंत या वर्तमान दिन तक कुल मिलाकर शेष",
         "previous-day": "पिछले दिन",
         "remove-entry": "प्रविष्टि निकालें",
-        "switch-view": "दृश्य स्विच करें",
-        "time-to-leave": "Time to Leave",
         "waived-day": "माफ किया हुआ दिन",
         "yes": "हाँ"
     },
@@ -112,77 +108,34 @@
         "add-waiver-day": "इस दिन के लिए एक छूट जोड़ें",
         "current-month": "वर्तमान माह पर जाएं",
         "day": "दिन",
-        "day-done-balance": "सब आज के लिए किया। दिन का संतुलन:",
         "entry-removal-confirmation": "क्या आप वाकई दिन से अंतिम दो प्रविष्टियाँ निकालना चाहते हैं",
         "last-day-balance": "अंतिम दिन का उपयोग संतुलन के लिए किया जाता है",
         "leave-by": "आज आप जिस समय पर आए हैं, उसके आधार पर आपको छोड़ देना चाहिए",
         "month-balance": "महीना शेष",
-        "month-balance-title": "इस महीने के लिए आज तक शेष राशि। एक सकारात्मक संतुलन का मतलब है अतिरिक्त घंटे जो आपको आज (या बाकी महीने) काम करने की आवश्यकता नहीं है।",
         "next-month": "अगले महीने",
         "no": "नहीं",
         "on": "पर",
-        "overall-balance": "समग्र संतुलन",
-        "overall-balance-title": "महीने के अंत या वर्तमान दिन तक कुल मिलाकर शेष",
         "previous-month": "पिछ्ला महिना",
         "remove-entry": "प्रविष्टि निकालें",
-        "switch-view": "दृश्य स्विच करें",
-        "time-to-leave": "Time to Leave",
         "total": "कुल",
         "waived-day": "माफ किया हुआ दिन",
         "working-days": "कार्य दिवस",
         "working-days-title": "महीने में कितने कार्य दिवस होते हैं",
         "yes": "हाँ"
     },
-    "$Calendar": {
-        "all-done": "सब आज के लिए किया। दिन का संतुलन:",
-        "current-month": "वर्तमान माह",
-        "current-month-title": "वर्तमान माह पर जाएं",
-        "day": "दिन",
-        "day-end": "डे एंड",
-        "day-start": "दिन की शुरुआत",
-        "day-total": "दिन का कुल",
-        "last-day-balance": "अंतिम दिन का उपयोग संतुलन के लिए किया जाता है",
-        "leave-by": "आज आप जिस समय पर आए हैं, उसके आधार पर आपको छोड़ देना चाहिए",
-        "lunch-begin": "दोपहर का भोजन शुरू",
-        "lunch-end": "दोपहर का भोजन",
-        "lunch-start": "दोपहर का खाना शुरू",
-        "lunch-total": "दोपहर का भोजन कुल",
+    "$BaseCalendar": {
+        "day-done-balance": "सब आज के लिए किया। दिन का संतुलन:",
         "month-balance": "महीना शेष",
         "month-balance-title": "इस महीने के लिए आज तक शेष राशि। एक सकारात्मक संतुलन का मतलब है अतिरिक्त घंटे जो आपको आज (या महीने के बाकी) काम करने की आवश्यकता नहीं है।",
-        "next-month": "अगले महीने",
-        "on": "पर",
         "overall-balance": "समग्र संतुलन",
         "overall-balance-title": "महीने के अंत या वर्तमान दिन तक कुल मिलाकर शेष",
-        "previous-month": "पिछ्ला महिना",
-        "punch-time": "पंच समय",
         "switch-view": "दृश्य स्विच करें",
-        "switch-view-title": "दृश्य स्विच करें",
-        "time-to-leave": "Time To Leave",
-        "working-days": "कार्य दिवस",
-        "working-days-title": "महीने में कितने कार्य दिवस होते हैं"
+        "time-to-leave": "Time To Leave"
     },
     "$Notification": {
         "punch-reminder": "में पंच करने के लिए मत भूलना!",
         "time-to-leave": "सुनो! मुझे लगता है कि यह छोड़ने का समय है।"
     },
-
-    "$FixedDayCalendar": {
-        "all-done": "सब आज के लिए किया। दिन का संतुलन:",
-        "day-end": "डे एंड",
-        "day-start": "दिन की शुरुआत",
-        "day-total": "दिन का कुल",
-        "leave-by": "आपको इसके द्वारा छोड़ देना चाहिए:",
-        "lunch-begin": "दोपहर का भोजन शुरू",
-        "lunch-end": "दोपहर का भोजन",
-        "lunch-total": "दोपहर का भोजन कुल",
-        "month-balance": "महीना शेष",
-        "month-balance-title": "इस महीने के लिए आज तक शेष राशि। एक सकारात्मक संतुलन का मतलब है अतिरिक्त घंटे जो आपको आज (या बाकी महीने) काम करने की आवश्यकता नहीं है।",
-        "non-working-day": "काम का दिन नहीं",
-        "overall-balance": "समग्र संतुलन",
-        "overall-balance-title": "महीने के अंत या वर्तमान दिन तक कुल मिलाकर शेष",
-        "waived-day": "माफ किया हुआ दिन"
-    },
-
     "$WorkdayWaiver": {
         "add-waiver": "छूट जोड़ें",
         "already-have-waiver": "आपके पास पहले से ही छूट है",
