+++ locales/id/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "Bersihkan Database",
         "confirm-clear-all-data": "Anda yakin Menghapus semua data?",
         "confirm-import-db": "Anda yakin ingin mengimpor database? ini akan menimpa informasi terkini.",
-        "confirm-migrate-to-flexible": "Anda yakin ingin memigrasi database kalender tetap ke kalender fleksibel? \n\nDatabase kalender fleksibel yang ada akan dihapus.",
         "copy": "Salin",
         "could-not-be-loaded": "Tidak dapat dimuat",
         "cut": "Pindah",
@@ -59,14 +58,16 @@
         "export": "Ekspor",
         "failed-entries": "Entri gagal",
         "failed-migrating": "Gagal migrasi",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "Bantuan",
         "import-database": "Impor database",
         "import-db-from-file": "Impor database dari file",
         "import-successful": "Yes! Impor berhasil!",
         "import": "Impor",
         "menu": "Menu",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "Migrasi kalender database tetap menjadi fleksibel",
-        "migrate-to-flexible-calendar": "Migrasi menjadi kalender fleksibel",
         "migration-successful": "Yay! Migrasi selesai!",
         "no-thanks": "Tidak, Terima Kasih",
         "ok": "OK",
@@ -76,9 +77,10 @@
         "punched-time": "Waktu direkam",
         "quit": "Keluar",
         "reload": "Memuat ulang",
+        "report": "Report",
         "select-all": "Pilih Semua",
         "send-feedback": "Berikan masukan",
-        "should-migrate-to-flexible": "Kalender fleksibel kamu kosong, Kamu ingin migrasi ke Kalender database tetap yang sudah ada?",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "Tampilkan Aplikasi",
         "something-went-wrong": "Suatu hal ada yang salah :(",
         "toggle-dev-tools": "Beralih ke alat pengembang",
@@ -88,17 +90,10 @@
         "yes-please": "Ya"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "Ubah tampilan",
         "current-day": "Hari terakhir",
         "previous-day": "Hari sebelumnya",
         "next-day": "Hari selanjutnya",
-        "time-to-leave": "Time to Leave",
         "leave-by": "Kamu seharusnya berangkat pada:",
-        "day-done-balance": "Semua selesai hari ini. Keseimbangan hari ini:",
-        "month-balance": "Keseimbangan bulan",
-        "month-balance-title": "Keseimbangan sampai hari ini untuk bulan ini. Keseimbangan positif berarti jam tambahan yang tidak Perlu Anda kerjakan hari ini (atau sisa bulan).",
-        "overall-balance": "Keseimbangan keseluruhan",
-        "overall-balance-title": "Keseimbangan keseluruhan sampai akhir bulan atau hari ini",
         "not-a-working-day": "Bukan hari kerja",
         "waived-day": "Hari pengecualian",
         "day-total": "Total hari",
@@ -109,23 +104,16 @@
         "entry": "Entri"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "Ubah tampilan",
         "current-month": "Bulan terakhir",
         "previous-month": "Bulan sebelumnya",
         "next-month": "Bulan selanjutnya",
-        "time-to-leave": "Time to Leave",
         "day": "Hari",
         "total": "Total",
         "leave-by": "Berdasarkan waktu Anda tiba hari ini, Anda harus berangkat",
-        "day-done-balance": "Semua selesai hari ini. Keseimbangan hari ini:",
         "last-day-balance": "Hari terakhir digunakan untuk keseimbangan",
         "on": "Hidup",
         "working-days": "Hari bekerja",
         "working-days-title": "Ada berapa hari kerja dalam sebulan",
-        "month-balance": "Keseimbangan bulan",
-        "month-balance-title": "Keseimbangan sampai hari ini untuk bulan ini. Keseimbangan positif berarti jam tambahan yang tidak Perlu Anda kerjakan hari ini (atau sisa bulan).",
-        "overall-balance": "Keseimbangan keseluruhan",
-        "overall-balance-title": "Keseimbangan keseluruhan sampai akhir bulan atau hari ini",
         "add-waiver-day": "Tambah hari pengecualian ",
         "waived-day": "Hari pengecualian",
         "remove-entry": "Hapus entri",
@@ -134,52 +122,17 @@
         "no": "Tidak"
     },
     "$Calendar": {
-        "leave-by": "Berdasarkan waktu Anda tiba hari ini, Anda harus berangkat",
-        "all-done": "Semua selesai hari ini. Keseimbangan hari ini:",
-        "last-day-balance": "Hari terakhir digunakan untuk keseimbangan",
-        "on": "Hidup",
-        "working-days": "Hari kerja",
-        "working-days-title": "Ada berapa hari kerja dalam sebulan",
+        "day-done-balance": "Semua selesai hari ini. Keseimbangan hari ini:",
         "month-balance": "Kesimbangan bulan",
         "month-balance-title": "Keseimbangan sampai hari ini untuk bulan ini. Keseimbangan positif berarti jam tambahan yang tidak Perlu Anda kerjakan hari ini (atau sisa bulan).",
         "overall-balance": "Keseimbangan keseluruhan",
         "overall-balance-title": "Keseimbangan keseluruhan sampai akhir bulan atau hari ini",
         "switch-view": "Ubah Tampilan",
-        "switch-view-title": "Ubah Tampilan",
-        "current-month": "Bulan Terakhir",
-        "current-month-title": "Bulan Terakhir",
-        "previous-month": "Bulan sebelumnya",
-        "next-month": "Bulan selanjutnya",
-        "time-to-leave": "Time to Leave",
-        "day-start": "Hari dimulai",
-        "lunch-begin": "Makan siang dimulai",
-        "lunch-end": "Makan siang berakhir",
-        "day-end": "Hari berakhir",
-        "lunch-total": "Total makan siang",
-        "day-total": "Total hari",
-        "day": "Hari",
-        "lunch-start": "Mulai makan siang",
-        "punch-time": "Rekam waktu"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "Hai. Saatnya kamu berangkat.",
         "punch-reminder": "Jangan lupa untuk rekam!"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "Kamu seharusnya berangkat pada:",
-        "all-done": "Semua selesai hari ini. Keseimbangan hari ini:",
-        "month-balance": "Kesimbangan bulan",
-        "month-balance-title": "Keseimbangan sampai hari ini untuk bulan ini. Keseimbangan positif berarti jam tambahan yang tidak Perlu Anda kerjakan hari ini (atau sisa bulan).",
-        "overall-balance": "Keseimbangan keseluruhan",
-        "overall-balance-title": "Keseimbangan keseluruhan sampai akhir bulan atau hari ini",
-        "non-working-day": "Bukan hari bekerja",
-        "waived-day": "Hari pengecualian",
-        "day-start": "Hari mulai",
-        "lunch-begin": "Makan siang dimulai",
-        "lunch-end": "Makan siang berakhir",
-        "day-end": "Hari berakhir",
-        "lunch-total": "Total makan siang",
-        "day-total": "Total hari"
     },
     "$WorkdayWaiver": {
         "add-waiver": "Tambahkan Pengecualian",
