+++ locales/it/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "Pulisci database",
         "confirm-clear-all-data": "Sei sicuro di voler pulire tutti i dati?",
         "confirm-import-db": "Sei sicuro di voler importare un database? Sovrascriverà tutte le informazioni attualmente presenti.",
-        "confirm-migrate-to-flexible": "Sei sicuro di voler migrare il database con calendario fisso sul calendario flessibile?\n\nIl database con calendario flessibile sarà cancellato.",
         "copy": "Copia",
         "could-not-be-loaded": "non può essere caricato",
         "cut": "Taglia",
@@ -59,14 +58,16 @@
         "export": "Esporta",
         "failed-entries": "Record falliti",
         "failed-migrating": "Fallimento migrando",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "Aiuto",
         "import-database": "Importa database",
         "import-db-from-file": "Importa database da file",
         "import-successful": "Yay! Importazione conclusa con successo!",
         "import": "Importa",
         "menu": "Menu",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "Migra database con calendario fisso a flessibile",
-        "migrate-to-flexible-calendar": "Migrate a calendar flessibile",
         "migration-successful": "Yay! Migrazione conclusa con successo!",
         "no-thanks": "No, grazie",
         "ok": "OK",
@@ -76,9 +77,10 @@
         "punched-time": "Orario di registrazione avvenuta",
         "quit": "Esci",
         "reload": "Ricarica",
+        "report": "Report",
         "select-all": "Seleziona tutto",
         "send-feedback": "Manda feedback",
-        "should-migrate-to-flexible": "Il tuo calendario flessibile è vuoto. Vuoi cominciare migrando il database con calendario fisso sul calendario flessibile?",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "Mostra App",
         "something-went-wrong": "Qualcosa è andato storto :(",
         "toggle-dev-tools": "Abilita Developer Tools",
@@ -88,17 +90,11 @@
         "yes-please": "Sì, per favore"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "Cambia visuale",
         "previous-day": "Giorno precedente",
         "next-day": "Giorno successivo",
-        "time-to-leave": "Time to Leave",
         "current-day": "Vai al giorno corrente",
-        "day-done-balance": "Tutto fatto per oggi. Bilancio del giorno:",
         "leave-by": "Dovresti andartene per le:",
-        "month-balance-title": "Bilancio del mese fino a oggi. Un bilancio positivo significa ore extra che non devi lavorare oggi (o nel resto del mese).",
         "month-balance": "Bilancio mensile",
-        "overall-balance-title": "Bilancio complessivo fino alla fine del mese o del giorno corrente",
-        "overall-balance": "Bilancio complessivo",
         "not-a-working-day": "Giorno non lavorativo",
         "waived-day": "Giorno escluso",
         "day-total": "Totale giornata",
@@ -109,23 +105,17 @@
         "entry": "Record"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "Cambia visuale",
         "current-month": "Vai al mese corrente",
         "previous-month": "Mese precedente",
         "next-month": "Prossimo mese",
-        "time-to-leave": "Time to Leave",
         "day": "Giorno",
         "total": "Totale",
         "leave-by": "In base all'ora di arrivo di oggi, dovresti andartene per le",
-        "day-done-balance": "Tutto fatto per oggi. Bilancio del giorno:",
         "last-day-balance": "Ultimo giorno considerato per il bilancio",
         "working-days-title": "Numero di giorni lavorativi del mese",
-        "month-balance-title": "Bilancio del mese fino a oggi. Un bilancio positivo significa ore extra che non devi lavorare oggi (o nel resto del mese).",
-        "overall-balance-title": "Bilancio complessivo fino alla fine del mese o del giorno corrente",
         "on": "Su",
         "working-days": "Giorni feriali",
         "month-balance": "Bilancio mensile",
-        "overall-balance": "Bilancio complessivo",
         "add-waiver-day": "Escludi questo giorno",
         "waived-day": "Giorno escluso",
         "remove-entry": "Rimuovi record",
@@ -133,53 +123,18 @@
         "yes": "Sì",
         "no": "No"
     },
-    "$Calendar": {
-        "leave-by": "In base all'ora di arrivo di oggi, dovresti andartene per le",
-        "all-done": "Tutto fatto per oggi. Bilancio del giorno:",
-        "last-day-balance": "Ultimo giorno considerato per il bilancio",
-        "on": "Su",
-        "working-days": "Giorni feriali",
-        "working-days-title": "Numero di giorni lavorativi del mese",
+    "$BaseCalendar": {
+        "day-done-balance": "Tutto fatto per oggi. Bilancio del giorno:",
         "month-balance": "Bilancio del mese",
         "month-balance-title": "Bilancio del mese fino a oggi. Un bilancio positivo significa ore extra che non devi lavorare oggi (o nel resto del mese).",
         "overall-balance": "Bilancio complessivo",
         "overall-balance-title": "Bilancio complessivo fino alla fine del mese o del giorno corrente",
         "switch-view": "Cambia visuale",
-        "switch-view-title": "Cambia visuale",
-        "current-month": "Mese corrente",
-        "current-month-title": "Vai al mese corrente",
-        "previous-month": "Mese precedente",
-        "next-month": "Mese successivo",
-        "time-to-leave": "Time to Leave",
-        "day-start": "Inizio giornata",
-        "lunch-begin": "Inizio pranzo",
-        "lunch-end": "Fine pranzo",
-        "day-end": "Fine giornata",
-        "lunch-total": "Totale pranzo",
-        "day-total": "Totale giornata",
-        "day": "Giorno",
-        "lunch-start": "Inizio pranzo",
-        "punch-time": "Ora di registrazione"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "Hey there! I think it's time to leave.",
         "punch-reminder": "Don't forget to punch in!"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "Dovresti andartene per le:",
-        "all-done": "Tutto fatto per oggi. Bilancio della giornata:",
-        "month-balance": "Bilancio del mese",
-        "month-balance-title": "Bilancio del mese fino a oggi. Un bilancio positivo significa ore extra che non devi lavorare oggi (o nel resto del mese).",
-        "overall-balance": "Bilancio complessivo",
-        "overall-balance-title": "Bilancio complessivo fino alla fine del mese o del giorno corrente",
-        "non-working-day": "Giorno non lavorativo",
-        "waived-day": "Giorno escluso",
-        "day-start": "Inizio giornata",
-        "lunch-begin": "Inizio pranzo",
-        "lunch-end": "Fine pranzo",
-        "day-end": "Fine giornata",
-        "lunch-total": "Totale pranzo",
-        "day-total": "Totale giornata"
     },
     "$WorkdayWaiver": {
         "add-waiver": "Aggiungi esclusione",
