+++ locales/ko/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "데이터베이스 삭제",
         "confirm-clear-all-data": "정말 모든 데이터를 삭제하시겠습니까?",
         "confirm-import-db": "정말 데이터베이스를 가져오시겠습니까? 현재 등록된 모든 정보를 덮어씌우게 됩니다.",
-        "confirm-migrate-to-flexible": "정말 고정 캘린더 데이터베이스를 가변 캘린더로 이전하시겠습니까?\n\n현재 가변 캘린더의 데이터베이스가 초기화됩니다.",
         "copy": "복사",
         "could-not-be-loaded": "로드할 수 없습니다",
         "cut": "잘라내기",
@@ -59,14 +58,16 @@
         "export": "내보내기",
         "failed-entries": "작업 내보내기 실패",
         "failed-migrating": "데이터 이전 실패",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "도움말",
         "import-database": "데이터베이스 가져오기",
         "import-db-from-file": "파일로부터 데이터베이스 가져오기",
         "import-successful": "와! 성공적으로 가져왔어요!",
         "import": "가져오기",
         "menu": "메뉴",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "고정 캘린더 데이터베이스를 가변으로 이전",
-        "migrate-to-flexible-calendar": "가변 캘린더로 데이터 이전",
         "migration-successful": "야호! 이전에 성공했어요!",
         "no-thanks": "아뇨, 괜찮아요",
         "ok": "확인",
@@ -76,9 +77,10 @@
         "punched-time": "기록된 시간",
         "quit": "나가기",
         "reload": "새로고침",
+        "report": "Report",
         "select-all": "모두 선택",
         "send-feedback": "피드백 보내기",
-        "should-migrate-to-flexible": "가변 캘린더가 비어 있습니다. 현재 고정 캘린더 데이터베이스를 가변 캘린더로 이전할까요?",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "앱 표시",
         "something-went-wrong": "뭔가 잘못된 것 같네요 :(",
         "toggle-dev-tools": "개발자 도구 토글",
@@ -88,17 +90,10 @@
         "yes-please": "네, 그래 주세요"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "보기 모드 전환",
         "current-day": "오늘로 이동",
         "previous-day": "이전 날",
         "next-day": "다음 날",
-        "time-to-leave": "Time to Leave",
         "leave-by": "이 시간에 퇴근할 수 있습니다:",
-        "day-done-balance": "오늘 작업이 모두 끝났습니다. 오늘의 작업 달성도:",
-        "month-balance": "이번 달 작업 달성도",
-        "month-balance-title": "현재까지의 이번 달 작업 달성 현황입니다. 양수 값은 목표 작업 시간보다 초과 근무하여 오늘(혹은 이번 달) 더 이상 일할 필요가 없다는 것을 나타냅니다.",
-        "overall-balance": "전체 작업 달성도",
-        "overall-balance-title": "월말 혹은 현재까지의 총 작업 달성 현황",
         "not-a-working-day": "오늘은 휴일입니다",
         "waived-day": "오늘은 휴가 날입니다",
         "day-total": "오늘의 총 작업 시간",
@@ -109,23 +104,16 @@
         "entry": "작업"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "보기 모드 전환",
         "current-month": "이번 달로 이동",
         "previous-month": "이전 달",
         "next-month": "다음 달",
-        "time-to-leave": "Time to Leave",
         "day": "일자",
         "total": "총 작업 시간",
         "leave-by": "오늘 출근 및 휴식 시간을 종합하면 이 시간에 퇴근할 수 있습니다:",
-        "day-done-balance": "오늘 작업이 모두 끝났습니다. 오늘의 작업 달성도:",
         "last-day-balance": "작업 달성도를 집계하는 마지막 일자",
         "on": "기준일",
         "working-days": "총 근무일",
         "working-days-title": "이번 달의 근무일 총합",
-        "month-balance": "이번 달 작업 달성도",
-        "month-balance-title": "현재까지의 이번 달 작업 달성 현황입니다. 양수 값은 목표 작업 시간보다 초과 근무하여 오늘(혹은 이번 달) 더 이상 일할 필요가 없다는 것을 나타냅니다.",
-        "overall-balance": "전체 작업 달성도",
-        "overall-balance-title": "월말 혹은 현재까지의 총 작업 달성 현황",
         "add-waiver-day": "이 일자에 휴가 등록",
         "waived-day": "휴가일",
         "remove-entry": "작업 단위 제거",
@@ -134,52 +122,17 @@
         "no": "아니오"
     },
     "$Calendar": {
-        "leave-by": "오늘 출근 및 휴식 시간을 종합하면 이 시간에 퇴근할 수 있습니다:",
-        "all-done": "오늘 작업이 모두 끝났습니다. 오늘의 작업 달성도:",
-        "last-day-balance": "작업 달성도를 집계하는 마지막 일자",
-        "on": "기준일",
-        "working-days": "총 근무일",
-        "working-days-title": "이번 달의 근무일 총합",
+        "day-done-balance": "오늘 작업이 모두 끝났습니다. 오늘의 작업 달성도:",
         "month-balance": "이번 달 작업 달성도",
         "month-balance-title": "현재까지의 이번 달 작업 달성 현황입니다. 양수 값은 목표 작업 시간보다 초과 근무하여 오늘(혹은 이번 달) 더 이상 일할 필요가 없다는 것을 나타냅니다.",
         "overall-balance": "전체 작업 달성도",
         "overall-balance-title": "월말 혹은 현재까지의 총 작업 달성 현황",
         "switch-view": "보기 모드 전환",
-        "switch-view-title": "보기 모드 전환",
-        "current-month": "이번 달",
-        "current-month-title": "이번 달로 이동",
-        "previous-month": "지난 달",
-        "next-month": "다음 달",
-        "time-to-leave": "Time to Leave",
-        "day-start": "작업 시작",
-        "lunch-begin": "점심 시작",
-        "lunch-end": "점심 종료",
-        "day-end": "작업 종료",
-        "lunch-total": "총 점심 시간",
-        "day-total": "총 작업 시간",
-        "day": "일자",
-        "lunch-start": "점심 시작",
-        "punch-time": "시간 기록"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "저기요! 이제 퇴근할 시간인 것 같은데요.",
         "punch-reminder": "시간 기록하는 것 잊지 마세요!"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "이 시간에 퇴근할 수 있습니다:",
-        "all-done": "오늘 작업이 모두 끝났습니다. 오늘의 작업 달성도:",
-        "month-balance": "이번 달 작업 달성도",
-        "month-balance-title": "현재까지의 이번 달 작업 달성 현황입니다. 양수 값은 목표 작업 시간보다 초과 근무하여 오늘(혹은 이번 달) 더 이상 일할 필요가 없다는 것을 나타냅니다.",
-        "overall-balance": "전체 작업 달성도",
-        "overall-balance-title": "월말 혹은 현재까지의 총 작업 달성 현황",
-        "non-working-day": "오늘은 휴일입니다",
-        "waived-day": "오늘은 휴가 날입니다",
-        "day-start": "작업 시작",
-        "lunch-begin": "점심 시작",
-        "lunch-end": "점심 종료",
-        "day-end": "작업 종료",
-        "lunch-total": "총 점심 시간",
-        "day-total": "총 작업 시간"
     },
     "$WorkdayWaiver": {
         "add-waiver": "휴가 추가",
