+++ locales/mr/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "डेटाबेस साफ करा",
         "confirm-clear-all-data": "आपली खात्री आहे की आपण सर्व डेटा साफ करू इच्छिता?",
         "confirm-import-db": "आपली खात्री आहे की आपण डेटाबेस आयात करू इच्छिता? हे कोणतीही सद्य माहिती अधिलिखित करेल.",
-        "confirm-migrate-to-flexible": "आपल्याला खात्री आहे की आपण निश्चित कॅलेंडरमध्ये निश्चित कॅलेंडर डेटाबेस स्थानांतरित करू इच्छिता?\n\nविद्यमान लवचिक कॅलेंडर डेटाबेस साफ केला जाईल.",
         "copy": "कॉपी करा",
         "could-not-be-loaded": "लोड केले जाऊ शकले नाही",
         "cut": "कट",
@@ -59,14 +58,16 @@
         "export-db-to-file": "फाईलमध्ये डेटाबेस निर्यात करा",
         "failed-entries": "नोंदी अयशस्वी",
         "failed-migrating": "स्थलांतर करण्यात अयशस्वी",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "मदत करा",
         "import": "आयात करा",
         "import-database": "आयात डेटाबेस",
         "import-db-from-file": "फाईलमधून डेटाबेस आयात करा",
         "import-successful": "होय! आयात यशस्वी!",
         "menu": "मेनू",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "निश्चित कॅलेंडर डेटाबेस लवचिकमध्ये स्थलांतरित करा",
-        "migrate-to-flexible-calendar": "लवचिक कॅलेंडरमध्ये स्थलांतर करा",
         "migration-successful": "होय! स्थलांतर यशस्वी!",
         "no-thanks": "नको धन्यवाद",
         "ok": "ठीक आहे",
@@ -76,9 +77,10 @@
         "punched-time": "मुक्कामी वेळ",
         "quit": "सोडा",
         "reload": "रीलोड करा",
+        "report": "Report",
         "select-all": "सर्व निवडा",
         "send-feedback": "अभिप्राय पाठवा",
-        "should-migrate-to-flexible": "आपले लवचिक कॅलेंडर रिक्त आहे. आपण आपल्या लवचिकमध्ये विद्यमान निश्चित कॅलेंडर डेटाबेसचे स्थानांतरन सुरू करू इच्छिता?",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "अ‍ॅप दर्शवा",
         "something-went-wrong": "काहीतरी चूक घडली :(",
         "toggle-dev-tools": "विकसक साधने टॉगल करा",
@@ -89,22 +91,16 @@
     },
     "$FlexibleDayCalendar": {
         "current-day": "चालू दिवस जा",
-        "day-done-balance": "आज सर्व केले. दिवसाची शिल्लक:",
         "day-total": "दिवस एकूण",
         "entry": "प्रवेश",
         "entry-removal-confirmation": "आपली खात्री आहे की आपण शेवटची प्रविष्टी पंक्ती हटवू इच्छिता?",
         "leave-by": "आपण याद्वारे सोडले पाहिजे:",
         "month-balance": "महिना शिल्लक",
-        "month-balance-title": "या महिन्यासाठी आज पर्यंत शिल्लक ठेवा. सकारात्मक संतुलनाचा अर्थ असा आहे की आपल्याला आज (किंवा महिना उर्वरित) काम करण्याची आवश्यकता नाही.",
         "next-day": "पुढचा दिवस",
         "no": "नाही",
         "not-a-working-day": "कामकाजाचा दिवस नाही",
-        "overall-balance": "एकूणच शिल्लक",
-        "overall-balance-title": "महिन्याच्या शेवटी किंवा चालू दिवसापर्यंत एकंदरीत शिल्लक",
         "previous-day": "आदल्या दिवशी",
         "remove-entry": "प्रविष्टी काढा",
-        "switch-view": "स्विच व्ह्यू",
-        "time-to-leave": "Time to Leave",
         "waived-day": "माफीचा दिवस",
         "yes": "होय"
     },
@@ -112,74 +108,33 @@
         "add-waiver-day": "या दिवसासाठी माफी जोडा",
         "current-month": "चालू महिन्यात जा",
         "day": "दिवस",
-        "day-done-balance": "आज सर्व केले. दिवसाची शिल्लक:",
         "entry-removal-confirmation": "आपल्याला खात्री आहे की आपण दिवसापासून शेवटच्या दोन प्रविष्ट्या काढू इच्छिता",
         "last-day-balance": "शिल्लक काम करण्यासाठी शेवटचा दिवस",
         "leave-by": "आपण आज पोहोचलेल्या वेळेच्या आधारे, आपण निघून जावे",
         "month-balance": "महिना शिल्लक",
-        "month-balance-title": "या महिन्यासाठी आज पर्यंत शिल्लक ठेवा. सकारात्मक संतुलनाचा अर्थ असा आहे की आपल्याला आज (किंवा महिना उर्वरित) काम करण्याची आवश्यकता नाही.",
         "next-month": "पुढील महिन्यात",
         "no": "नाही",
         "on": "चालू",
-        "overall-balance": "एकूणच शिल्लक",
-        "overall-balance-title": "महिन्याच्या शेवटी किंवा चालू दिवसापर्यंत एकंदरीत शिल्लक",
         "previous-month": "मागील महिना",
         "remove-entry": "प्रविष्टी काढा",
-        "switch-view": "स्विच व्ह्यू",
-        "time-to-leave": "Time to Leave",
         "total": "एकूण",
         "waived-day": "माफीचा दिवस",
         "working-days": "कामाचे दिवस",
         "working-days-title": "महिन्यात किती कार्य दिवस आहेत",
         "yes": "होय"
     },
-    "$Calendar": {
-        "all-done": "आज सर्व केले. दिवसाची शिल्लक:",
-        "current-month": "चालू महिना",
-        "current-month-title": "चालू महिन्यात जा",
-        "day": "दिवस",
-        "day-end": "डे एंड",
-        "day-start": "दिवस प्रारंभ",
-        "day-total": "दिवस एकूण",
-        "last-day-balance": "शिल्लक काम करण्यासाठी शेवटचा दिवस",
-        "leave-by": "आपण आज पोहोचलेल्या वेळेच्या आधारे, आपण निघून जावे",
-        "lunch-begin": "दुपारचे जेवण",
-        "lunch-end": "लंच एंड",
-        "lunch-start": "लंच स्टार्ट",
-        "lunch-total": "लंच टोटल",
+    "$BaseCalendar": {
+        "day-done-balance": "आज सर्व केले. दिवसाची शिल्लक:",
         "month-balance": "महिना शिल्लक",
         "month-balance-title": "या महिन्यासाठी आज पर्यंत शिल्लक ठेवा. सकारात्मक संतुलनाचा अर्थ असा आहे की आपल्याला आज (किंवा महिना उर्वरित) काम करण्याची आवश्यकता नाही.",
-        "next-month": "पुढील महिन्यात",
-        "on": "चालू",
         "overall-balance": "एकूणच शिल्लक",
         "overall-balance-title": "महिन्याच्या शेवटी किंवा चालू दिवसापर्यंत एकंदरीत शिल्लक",
-        "previous-month": "मागील महिना",
-        "punch-time": "पंच वेळ",
         "switch-view": "स्विच व्ह्यू",
-        "switch-view-title": "स्विच व्ह्यू",
-        "time-to-leave": "Time to Leave",
-        "working-days": "कामाचे दिवस",
-        "working-days-title": "महिन्यात किती कार्य दिवस आहेत"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "punch-reminder": "मध्ये मुक्काम विसरू नका!",
         "time-to-leave": "अहो तिथे! मला वाटते आता निघण्याची वेळ आली आहे."
-    },
-    "$FixedDayCalendar": {
-        "all-done": "आज सर्व केले. दिवसाची शिल्लक:",
-        "day-end": "डे एंड",
-        "day-start": "दिवस प्रारंभ",
-        "day-total": "दिवस एकूण",
-        "leave-by": "आपण याद्वारे सोडले पाहिजे:",
-        "lunch-begin": "दुपारचे जेवण",
-        "lunch-end": "लंच एंड",
-        "lunch-total": "लंच टोटल",
-        "month-balance": "महिना शिल्लक",
-        "month-balance-title": "या महिन्यासाठी आज पर्यंत शिल्लक ठेवा. सकारात्मक संतुलनाचा अर्थ असा आहे की आपल्याला आज (किंवा महिना उर्वरित) काम करण्याची आवश्यकता नाही.",
-        "non-working-day": "कामकाजाचा दिवस नाही",
-        "overall-balance": "एकंदरीत शिल्लक",
-        "overall-balance-title": "महिन्याच्या शेवटी किंवा चालू दिवसापर्यंत एकंदरीत शिल्लक",
-        "waived-day": "माफीचा दिवस"
     },
     "$WorkdayWaiver": {
         "add-waiver": "माफी जोडा",
