+++ locales/nl/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "Database leegmaken",
         "confirm-clear-all-data": "Weet je zeker dat je alle gegevens wilt wissen?",
         "confirm-import-db": "Weet je zeker dat je een database wilt inladen? Het zal alle huidige informatie overschrijven.",
-        "confirm-migrate-to-flexible": "Weet je zeker dat je de vaste kalenderdatabase wilt overzetten naar de flexibele kalender?\n\nDe bestaande flexibele kalenderdatabase zal worden gewist.",
         "copy": "Kopiëren",
         "could-not-be-loaded": "kon niet worden geladen",
         "cut": "Knippen",
@@ -59,14 +58,16 @@
         "export": "Opslaan",
         "failed-entries": "Mislukte invoeringen",
         "failed-migrating": "Overzetten mislukt",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "Help",
         "import-database": "Database inladen",
         "import-db-from-file": "Databasebestand inladen",
         "import-successful": "Joepie! Succesvol ingeladen!",
         "import": "Inladen",
         "menu": "Menu",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "De vaste kalenderdatabase naar een flexibele overzetten",
-        "migrate-to-flexible-calendar": "Naar een flexibele kalender overzetten",
         "migration-successful": "Joepie! Succesvol overgezet!",
         "no-thanks": "Nee, bedankt",
         "ok": "Oké",
@@ -76,9 +77,10 @@
         "punched-time": "Ingevoerde tijd",
         "quit": "Sluiten",
         "reload": "Herladen",
+        "report": "Report",
         "select-all": "Alles selecteren",
         "send-feedback": "Feedback versturen",
-        "should-migrate-to-flexible": "Je flexibele kalender is leeg. Wil je beginnen met het overzetten van de bestaande vaste kalenderdatabase naar je flexibele kalender?",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "App bekijken",
         "something-went-wrong": "Er is iets misgegaan :(",
         "toggle-dev-tools": "Ontwikkelaarshulpmiddelen in-/uitschakelen",
@@ -88,17 +90,10 @@
         "yes-please": "Ja, graag"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "Weergave veranderen",
         "current-day": "Naar de huidige dag gaan",
         "previous-day": "Vorige dag",
         "next-day": "Volgende dag",
-        "time-to-leave": "Time to Leave",
         "leave-by": "Het is tijd om te gaan om:",
-        "day-done-balance": "Het is tijd om te gaan. Overzicht van vandaag:",
-        "month-balance": "Maandelijks overzicht",
-        "month-balance-title": "Overzicht tot vandaag voor deze maand. Een positief overzicht betekent dat je vandaag (of de rest van de maand) geen extra uren hoeft te werken.",
-        "overall-balance": "Algemeen overzicht",
-        "overall-balance-title": "Algemeen overzicht tot het einde van de maand of de huidige dag",
         "not-a-working-day": "Geen werkdag",
         "waived-day": "Vrije dag",
         "day-total": "Dagtotaal",
@@ -109,23 +104,16 @@
         "entry": "Invoer"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "Weergave veranderen",
         "current-month": "Naar de huidige maand gaan",
         "previous-month": "Vorige maand",
         "next-month": "Volgende maand",
-        "time-to-leave": "Time to Leave",
         "day": "Dag",
         "total": "Totaal",
         "leave-by": "Gebaseerd op de tijd dat je vandaag begon, zou je moeten stoppen om",
-        "day-done-balance": "Klaar voor vandaag. Overzicht van vandaag:",
         "last-day-balance": "Laatste dag gebruikt voor overzicht",
         "on": "Op",
         "working-days": "Werkdagen",
         "working-days-title": "Hoeveel werkdagen er in de maand zijn",
-        "month-balance": "Maandelijks overzicht",
-        "month-balance-title": "Overzicht tot vandaag voor deze maand. Een positief overzicht betekent dat je vandaag (of de rest van de maand) geen extra uren hoeft te werken.",
-        "overall-balance": "Algemeen overzicht",
-        "overall-balance-title": "Algemeen overzicht tot het einde van de maand of de huidige dag",
         "add-waiver-day": "Voeg een uitzondering toe voor deze dag",
         "waived-day": "Vrije dag",
         "remove-entry": "Gegevens verwijderen",
@@ -134,52 +122,17 @@
         "no": "Nee"
     },
     "$Calendar": {
-        "leave-by": "Gebaseerd op de tijd dat je vandaag begon, zou je moeten stoppen om",
-        "all-done": "Klaar voor vandaag. Overzicht van vandaag:",
-        "last-day-balance": "Laatste dag gebruikt voor overzicht",
-        "on": "Op",
-        "working-days": "Werkdagen",
-        "working-days-title": "Hoeveel werkdagen er in de maand zijn",
+        "day-done-balance": "Het is tijd om te gaan. Overzicht van vandaag:",
         "month-balance": "Maandelijks overzicht",
         "month-balance-title": "Overzicht tot vandaag voor deze maand. Een positief overzicht betekent dat je vandaag (of de rest van de maand) geen extra uren hoeft te werken.",
         "overall-balance": "Algemeen overzicht",
         "overall-balance-title": "Algemeen overzicht tot het einde van de maand of de huidige dag",
         "switch-view": "Weergave veranderen",
-        "switch-view-title": "Weergave veranderen",
-        "current-month": "Huidige maand",
-        "current-month-title": "Ga naar de huidige maand",
-        "previous-month": "Vorige maand",
-        "next-month": "Volgende maand",
-        "time-to-leave": "Time to Leave",
-        "day-start": "Dag begin",
-        "lunch-begin": "Pauze begin",
-        "lunch-end": "Pauze einde",
-        "day-end": "Dag einde",
-        "lunch-total": "Pauze totaal",
-        "day-total": "Dag totaal",
-        "day": "Dag",
-        "lunch-start": "Pauze begin",
-        "punch-time": "Huidige tijd invoeren"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "Hé daar! Ik denk dat het tijd is om te gaan.",
         "punch-reminder": "Vergeet niet de begintijd in te voeren!"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "Het is tijd om te gaan om:",
-        "all-done": "Klaar voor vandaag. Overzicht van vandaag:",
-        "month-balance": "Maandelijks overzicht",
-        "month-balance-title": "Overzicht tot vandaag voor deze maand. Een positief overzicht betekent dat je vandaag (of de rest van de maand) geen extra uren hoeft te werken.",
-        "overall-balance": "Algemeen overzicht",
-        "overall-balance-title": "Algemeen overzicht tot het einde van de maand of de huidige dag",
-        "non-working-day": "Geen werkdag",
-        "waived-day": "Vrije dag",
-        "day-start": "Dag begin",
-        "lunch-begin": "Pauze begin",
-        "lunch-end": "Pauze einde",
-        "day-end": "Dag einde",
-        "lunch-total": "Pauze totaal",
-        "day-total": "Dag totaal"
     },
     "$WorkdayWaiver": {
         "add-waiver": "Vrijstellingen",
