+++ locales/pl/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "Wyczyść dane",
         "confirm-clear-all-data": "Czy jesteś pewien, że chcesz usunąć wszystkie dane?",
         "confirm-import-db": "Czy jesteś pewien, że chcesz wczytać dane? Zastąpią one istniejące.",
-        "confirm-migrate-to-flexible": "Czy jesteś pewien, że chesz przenieść dane stałego kalendarza do dynamicznego kalendarza?\n\nJego dane zostaną wyczyszczone.",
         "copy": "Kopiuj",
         "could-not-be-loaded": "nie może być załadowany.",
         "cut": "Wytnij",
@@ -59,14 +58,16 @@
         "export": "Wyeksportuj",
         "failed-entries": "Failed entries",
         "failed-migrating": "Przenoszenie nie powiodło się",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "Pomoc",
         "import-database": "Importuj bazę danych",
         "import-db-from-file": "Importuj bazę danych z pliku",
         "import-successful": "Jej! Pomyślnie zaimportowano!",
         "import": "Importuj",
         "menu": "Menu",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "Przenieś dane stałego kalendarza do dynamicznego",
-        "migrate-to-flexible-calendar": "Przenieś kalendarz dynamiczny",
         "migration-successful": "Jej! Przenoszenie udane!",
         "no-thanks": "Nie, dziękuję",
         "ok": "Okej",
@@ -76,9 +77,10 @@
         "punched-time": "Rozpoczęto odlicznie",
         "quit": "Wyjdź",
         "reload": "Odśwież",
+        "report": "Report",
         "select-all": "Wybierz wszystkie",
         "send-feedback": "Prześlij opinię",
-        "should-migrate-to-flexible": "Twój dynamiczny kalendarz jest pusty. Chcesz do niego przenieść dane stałego kalendarza?",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "Pokaż aplikację",
         "something-went-wrong": "Coś poszło nie tak :(",
         "toggle-dev-tools": "Przełącz narzędzia programistyczne",
@@ -88,17 +90,10 @@
         "yes-please": "Tak, poproszę"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "Zmień widok",
         "current-day": "Pokaż dziś",
         "previous-day": "Poprzedni dzień",
         "next-day": "Następny dzień",
-        "time-to-leave": "Time to Leave",
         "leave-by": "Powinieneś skończyć pracę o:",
-        "day-done-balance": "To wszystko na dzisiaj. Bilans dnia: ",
-        "month-balance": "Bilans miesięczny",
-        "month-balance-title": "Bilans czasu pracy od początku miesiąca do dzisiaj. Dodatni wynik oznacza nadgodziny nie musisz już więcej dzisiaj pracować (ani do końca miesiąca).",
-        "overall-balance": "Bilans całkowity",
-        "overall-balance-title": "Bilans czasu pracy do tej pory",
         "not-a-working-day": "Wolne tego dnia!",
         "waived-day": "Zwolnienie",
         "day-total": "Dzisiejszy czas pracy:",
@@ -109,23 +104,16 @@
         "entry": "Wpis"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "Zmień widok",
         "current-month": "Pokaż obecny miesiąc",
         "previous-month": "Poprzedni miesiąc",
         "next-month": "Następny miesiąc",
-        "time-to-leave": "Time to Leave",
         "day": "Dzień",
         "total": "Suma",
         "leave-by": "Zgodnie z czasem rozpoczęcia, powinieneś zakończyć pracę o:",
-        "day-done-balance": "To wszystko na dzisiaj. Bilans dnia: ",
         "last-day-balance": "Ostatni dzień użyty do bilansu",
         "on": "Ostatnio",
         "working-days": "Dni pracy",
         "working-days-title": "Ile jest dni pracy w miesiącu",
-        "month-balance": "Bilans miesięczny",
-        "month-balance-title": "Bilans czasu pracy od początku miesiąca do dzisiaj. Dodatni wynik oznacza nadgodziny nie musisz już więcej dzisiaj pracować (ani do końca miesiąca).",
-        "overall-balance": "Bilans całkowity",
-        "overall-balance-title": "Bilans czasu pracy do tej pory",
         "add-waiver-day": "Dodaj zwolnienie tego dnia",
         "waived-day": "Dzień zwolnienia",
         "remove-entry": "Usuń wpis",
@@ -134,52 +122,17 @@
         "no": "Nie"
     },
     "$Calendar": {
-        "leave-by": "Zgodnie z czasem rozpoczęcia, powinieneś zakończyć pracę o:",
-        "all-done": "To wszystko na dzisiaj. Bilans dnia: ",
-        "last-day-balance": "Ostatni dzień użyty do bilansu",
-        "on": "Ostatnio",
-        "working-days": "Dni pracy",
-        "working-days-title": "Ile jest dni pracy w miesiącu",
+        "day-done-balance": "To wszystko na dzisiaj. Bilans dnia:",
         "month-balance": "Bilans miesięczny",
         "month-balance-title": "Bilans czasu pracy od początku miesiąca do dzisiaj. Dodatni wynik oznacza nadgodziny nie musisz już więcej dzisiaj pracować (ani do końca miesiąca).",
         "overall-balance": "Bilans całkowity",
         "overall-balance-title": "Bilans czasu pracy do tej pory",
         "switch-view": "Zmień widok",
-        "switch-view-title": "Widok dnia/miesiąca",
-        "current-month": "Obecny miesiąc",
-        "current-month-title": "Pokaż obecny miesiąc",
-        "previous-month": "Poprzedni miesiąc",
-        "next-month": "Następny miesiąc",
-        "time-to-leave": "Time to Leave",
-        "day-start": "Początek dnia",
-        "lunch-begin": "Początek lanch'u",
-        "lunch-end": "Koniec lanch'u",
-        "day-end": "Koniec dnia",
-        "lunch-total": "Czas lanch'u",
-        "day-total": "Czas pracy",
-        "day": "Dzień",
-        "lunch-start": "Początek lunch'u",
-        "punch-time": "Zapisz czas"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "Hej, to chyba już czas by skończyć pracę i się odprężyć!",
         "punch-reminder": "Nie zapomnij podbić czasu pracy!"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "Powinieneś skończyć pracę o:",
-        "all-done": "To wszystko na dzisiaj. Bilans dnia: ",
-        "month-balance": "Bilans miesięczny",
-        "month-balance-title": "Bilans czasu pracy od początku miesiąca do dzisiaj. Dodatni wynik oznacza nadgodziny nie musisz już więcej dzisiaj pracować (ani do końca miesiąca).",
-        "overall-balance": "Bilans całkowity",
-        "overall-balance-title": "Bilans czasu pracy do tej pory",
-        "non-working-day": "Ten dzień jest wolny",
-        "waived-day": "Dzień zwolnienia",
-        "day-start": "Początek dnia",
-        "lunch-begin": "Początek lanch'u",
-        "lunch-end": "Koniec lanch'u",
-        "day-end": "Koniec dnia",
-        "lunch-total": "Czas lanch'u",
-        "day-total": "Czas pracy"
     },
     "$WorkdayWaiver": {
         "add-waiver": "Dodaj zwolnienie",
