+++ locales/pt-BR/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "Limpar base de dados",
         "confirm-clear-all-data": "Tem certeza que quer limpar todos os dados?",
         "confirm-import-db": "Tem certeza que deseja importar uma base de dados? Ela vai sobrescrever qualquer informação corrente.",
-        "confirm-migrate-to-flexible": "Tem certeza que deseja migrar da base de dados do calendário fixo para a do flexível?\n\nA base de dados do calendário flexível existente vai ser limpa.",
         "copy": "Copiar",
         "could-not-be-loaded": "não foram carregadas",
         "cut": "Cortar",
@@ -59,14 +58,16 @@
         "export": "Exportar",
         "failed-entries": "Entradas falharam",
         "failed-migrating": "Migração falhou",
+        "fresh-start": "Começar do zero",
+        "fresh-start-confirm": "Tem certeza que quer começar do zero?",
         "help": "Ajuda",
         "import-database": "Importar base de dados",
         "import-db-from-file": "Importar Base de Dados de um Arquivo",
         "import-successful": "Uhul! Sucesso ao importar!",
         "import": "Importar",
         "menu": "Menu",
+        "migrate": "Migrar",
         "migrate-calendar-to-flexible": "Migrar base de dados de calendário fixo para flexível",
-        "migrate-to-flexible-calendar": "Migrar para calendário flexível",
         "migration-successful": "Uhul! Sucesso na migração!",
         "no-thanks": "Não, valeu",
         "ok": "OK",
@@ -76,9 +77,10 @@
         "punched-time": "Ponto registrado",
         "quit": "Fechar",
         "reload": "Recarregar",
+        "report": "Reportar",
         "select-all": "Selecionar Tudo",
         "send-feedback": "Enviar comentários",
-        "should-migrate-to-flexible": "Seu calendário flexível está vazio. Você quer começar migrando o calendário fixo existente para o flexível?",
+        "should-migrate-to-flexible": "Um dos novos recursos deste lançamento do Time to Leave é um número flexível de entradas para o calendário.\n\nEssa nova visão usa um novo formato de banco de dados.\nPara usá-lo, você pode ou migrar sua base de dados existente para o novo formato, ou começar do zero.",
         "show-app": "Mostrar App",
         "something-went-wrong": "Algo de errado aconteceu :(",
         "toggle-dev-tools": "Alternar DevTools",
@@ -88,17 +90,11 @@
         "yes-please": "Sim, por favor"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "Trocar Visualização",
         "current-day": "Ir para o Dia Atual",
         "previous-day": "Dia Anterior",
         "next-day": "Próximo Dia",
-        "time-to-leave": "Time to Leave",
         "leave-by": "Você deve sair às:",
-        "day-done-balance": "Tudo certo por hoje. Saldo do dia:",
         "month-balance": "Saldo Mensal",
-        "month-balance-title": "Saldo até hoje para este mês. Um saldo positivo significa horas extras que você não precisa trabalhar hoje (ou o resto do mês)",
-        "overall-balance": "Saldo Total",
-        "overall-balance-title": "Saldo total até o fim do mês ou dia atual",
         "not-a-working-day": "Nao é um dia útil",
         "waived-day": "Dia de folga",
         "day-total": "Total do Dia",
@@ -109,23 +105,17 @@
         "entry": "Registro"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "Trocar Visualização",
         "current-month": "Ir para o Mês Atual",
         "previous-month": "Mês Anterior",
         "next-month": "Próximo Mês",
-        "time-to-leave": "Time to Leave",
         "day": "Dia",
         "total": "Total",
         "leave-by": "Com base na hora em que você chegou hoje, você deve sair às",
-        "day-done-balance": "Tudo certo por hoje. Saldo do dia:",
         "last-day-balance": "Último dia usado para o saldo",
         "on": "Em",
         "working-days": "Dias úteis",
         "working-days-title": "Quantos dias úteis existem no mês",
         "month-balance": "Saldo do Mês",
-        "month-balance-title": "Saldo até hoje para este mês. Um saldo positivo significa horas extras em que você não precisa trabalhar hoje (ou no restante do mês).",
-        "overall-balance": "Saldo Total",
-        "overall-balance-title": "Saldo total até o fim do mês ou dia atual",
         "add-waiver-day": "Adicionar dia de folga",
         "waived-day": "Dia de Folga",
         "remove-entry": "Remover registro",
@@ -133,53 +123,18 @@
         "yes": "Sim",
         "no": "Não"
     },
-    "$Calendar": {
-        "leave-by": "Com base na hora em que você chegou hoje, você deve sair às",
-        "all-done": "Tudo certo por hoje. Saldo do dia:",
-        "last-day-balance": "Último dia usado para o saldo",
-        "on": "Em",
-        "working-days": "Dias úteis",
-        "working-days-title": "Quantos dias úteis existem no mês",
+    "$BaseCalendar": {
+        "day-done-balance": "Tudo certo por hoje. Saldo do dia:",
         "month-balance": "Saldo do Mês",
         "month-balance-title": "Saldo até hoje para este mês. Um saldo positivo significa horas extras em que você não precisa trabalhar hoje (ou no restante do mês).",
         "overall-balance": "Saldo Total",
         "overall-balance-title": "Saldo total até o fim do mês ou dia atual",
         "switch-view": "Trocar Visualização",
-        "switch-view-title": "Trocar Visualização",
-        "current-month": "Mês Atual",
-        "current-month-title": "Vá para Mês Atual",
-        "previous-month": "Mês Anterior",
-        "next-month": "Próximo Mês",
-        "time-to-leave": "Time to Leave",
-        "day-start": "Início do Dia",
-        "lunch-begin": "Início do Almoço",
-        "lunch-end": "Fim do Almoço",
-        "day-end": "Fim do Dia",
-        "lunch-total": "Total do Almoço",
-        "day-total": "Total do Dia",
-        "day": "Dia",
-        "lunch-start": "Início do Almoço",
-        "punch-time": "Bater ponto"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "Ei você! Acho que é hora de ir.",
         "punch-reminder": "Não esqueça de bater ponto!"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "Você deve sair às:",
-        "all-done": "Tudo certo por hoje. Saldo do dia:",
-        "month-balance": "Saldo do Mês",
-        "month-balance-title": "Saldo até hoje para este mês. Um saldo positivo significa horas extras em que você não precisa trabalhar hoje (ou no restante do mês).",
-        "overall-balance": "Saldo Total",
-        "overall-balance-title": "Saldo total até o fim do mês ou dia atual",
-        "non-working-day": "Dia não útil",
-        "waived-day": "Dia de Folga",
-        "day-start": "Início do Dia",
-        "lunch-begin": "Início do Almoço",
-        "lunch-end": "Fim do Almoço",
-        "day-end": "Fim do Dia",
-        "lunch-total": "Total do Almoço",
-        "day-total": "Total do Dia"
     },
     "$WorkdayWaiver": {
         "add-waiver": "Adicionar folga",
