+++ locales/th-TH/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "ลบข้อมูลในฐานข้อมูล",
         "confirm-clear-all-data": "ยืนยันการลบข้อมูลทั้งหมด",
         "confirm-import-db": "ยืนยันการนำเข้าฐานข้อมูล ข้อมูลที่ถูกนำเข้าจะแทนที่ข้อมูลในปัจจุบัน",
-        "confirm-migrate-to-flexible": "คุณมั่นใจหรือไม่ที่จะย้ายฐานข้อมูลปฏิทินแบบล็อคไม่ยังปฏิทินแบบยืดหยุ่น?\n\nปฏิทินแบบยืดหยุ่นที่มีอยู่จะถูกล้างข้อมูล",
         "copy": "คัดลอก",
         "could-not-be-loaded": "ไม่สามารถโหลดข้อมูลได้",
         "cut": "ตัด",
@@ -59,14 +58,16 @@
         "export": "นำออก",
         "failed-entries": "ข้อมูลที่ล้มเหลว",
         "failed-migrating": "การย้ายข้อมูลล้มเหลว",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "ช่วยเหลือ",
         "import-database": "นำเข้าฐานข้อมูล",
         "import-db-from-file": "การนำเข้าฐานข้อมูลจากไฟล์",
         "import-successful": "การนำเข้าฐานข้อมูลเสร็จสิ้น",
         "import": "นำเข้า",
         "menu": "เมนู",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "ย้ายไปยังปฎิทินแบบล็อคไปยังแบบยืดหยุ่น",
-        "migrate-to-flexible-calendar": "ย้ายไปยังปฎิทินแบบยืดหยุ่น",
         "migration-successful": "เย้! การโอนย้ายเสร็จสมบูรณ์ !",
         "no-thanks": "ไม่",
         "ok": "ตกลง",
@@ -76,9 +77,10 @@
         "punched-time": "เวลาที่ถูกบันทึก",
         "quit": "ออกจากโปรแกรม",
         "reload": "โหลดอีกครั้ง",
+        "report": "Report",
         "select-all": "เลือกทั้งหมด",
         "send-feedback": "ส่งคำติชม",
-        "should-migrate-to-flexible": "ไม่พบปฎิทินยืดหยุ่นของคุณ. คุณต่อการโอนย้ายฐานข้อมูลปฎิทินแบบล็อคไปยังแบบยืดหยุ่นหรือไม่?",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "เปิดโปรแกรม",
         "something-went-wrong": "เกิดข้อผิดพลาด :(",
         "toggle-dev-tools": "เปิดมุมมองนักพัฒนา",
@@ -88,17 +90,10 @@
         "yes-please": "ใช่"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "เปลี่ยนมุมมอง",
         "current-day": "ไปที่วันปัจจุบัน",
         "previous-day": "เมื่อวานนี้",
         "next-day": "พรุ่งนี้",
-        "time-to-leave": "Time to Leave",
         "leave-by": "ควรเลิกงาน:",
-        "day-done-balance": "วันนี้เรียบร้อยแล้ว จำนวนชั่วโมงส่วนเกินของวันนี้:",
-        "month-balance": "จำนวนชั่วโมงส่วนเกินในเดือนนี้",
-        "month-balance-title": "จำนวนชั่วโมงส่วนเกินจนถึงวันนี้สำหรับเดือนนี้ จำนวนที่เป็นบวกหมายถึงชั่วโมงส่วนเกินที่คุณไม่ต้องทำงานสำหรับวันนี้ (หรือสำหรับวันที่เหลือของเดือน)",
-        "overall-balance": "จำนวนชั่วโมงส่วนเกิน",
-        "overall-balance-title": "จำนวนช่วงโมงส่วนเกินจนถึงสิ้นเดือนหรือวันนี้",
         "not-a-working-day": "ไม่ใช่วันทำงาน",
         "waived-day": "วันหยุด",
         "day-total": "วันทั้งหมด",
@@ -109,23 +104,16 @@
         "entry": "รายการ"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "เปลี่ยนมุมมอง",
         "current-month": "ไปยังเดือนปัจจุบัน",
         "previous-month": "เดือนที่แล้ว",
         "next-month": "เดือนหน้า",
-        "time-to-leave": "Time to Leave",
         "day": "วัน",
         "total": "ทั้งหมด",
         "leave-by": "จากเวลาที่เข้างานวันนี้, คุณควรเลิกงานเวลา",
-        "day-done-balance": "วันนี้เรียบร้อยแล้ว จำนวนชั่วโมงส่วนเกินของวันนี้:",
         "last-day-balance": "วันสุดท้ายที่ได้ใช้ชั่วโมงส่วนเกิน",
         "on": "ในวัน",
         "working-days": "วันทำงาน",
         "working-days-title": "จำนวนวันทำงานในเดือน",
-        "month-balance": "จำนวนชั่วโมงส่วนเกินในเดือนนี้",
-        "month-balance-title": "จำนวนชั่วโมงส่วนเกินจนถึงวันนี้สำหรับเดือนนี้ จำนวนที่เป็นบวกหมายถึงชั่วโมงส่วนเกินที่คุณไม่ต้องทำงานสำหรับวันนี้ (หรือสำหรับวันที่เหลือของเดือน)",
-        "overall-balance": "จำนวนชั่วโมงส่วนเกินทั้งหมด",
-        "overall-balance-title": "จำนวนชั่วโมงส่วนเกินทั้งหมด นับถึงวันปัจจุบันหรือสิ้นเดือนนี้",
         "add-waiver-day": "เพิ่มวันหยุดสำหรับวันนี้",
         "waived-day": "วันหยุด",
         "remove-entry": "ลบรายการ",
@@ -134,52 +122,17 @@
         "no": "ไม่"
     },
     "$Calendar": {
-        "leave-by": "วันนี้ควรเลิกงาน",
-        "all-done": "วันนี้จบแล้ว. จำนวนชั่วโมงส่วนเกินของวันนี้:",
-        "last-day-balance": "วันสุดท้ายที่ได้ใช้ชั่วโมงส่วนเกิน",
-        "on": "เมื่อวันที่",
-        "working-days": "จำนวนวันทำงาน",
-        "working-days-title": "จำนวนวันทำงานในแต่ละเดือน",
+        "day-done-balance": "วันนี้เรียบร้อยแล้ว จำนวนชั่วโมงส่วนเกินของวันนี้:",
         "month-balance": "จำนวนชั่วโมงส่วนเกินในเดือนนี้",
         "month-balance-title": "จำนวนชั่วโมงส่วนเกินในเดือนนี้ ถ้าค่าเป็นบวกแสดงว่าทำเกินมา ถ้าค่าติดลบแสดงว่าทำขาดไป",
         "overall-balance": "จำนวนชั่วโมงส่วนเกินทั้งหมด",
         "overall-balance-title": "จำนวนชั่วโมงส่วนเกินทั้งหมด นับถึงวันปัจจุบันหรือสิ้นเดือนนี้",
         "switch-view": "เปลี่ยนมุมมอง",
-        "switch-view-title": "เปลี่ยนมุมมอง",
-        "current-month": "เดือนปัจจุบัน",
-        "current-month-title": "ไปที่เดือนปัจจุบัน",
-        "previous-month": "เดือนที่แล้ว",
-        "next-month": "เดือนถัดไป",
-        "time-to-leave": "Time to Leave",
-        "day-start": "เริ่มทำงาน",
-        "lunch-begin": "เริ่มพักเที่ยง",
-        "lunch-end": "สิ้นสุดพักเที่ยง",
-        "day-end": "สิ้นสุดการทำงาน",
-        "lunch-total": "รวมเวลาพักเที่ยง",
-        "day-total": "รวมเวลาทำงาน",
-        "day": "วันที่",
-        "lunch-start": "เริ่มพักเที่ยง",
-        "punch-time": "บันทึกเวลา"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "เลิกงานได้แล้ว 🏡.",
         "punch-reminder": "อย่าลืมบันทึกเวลานะ! 🧐"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "ควรเลิกงานตอน:",
-        "all-done": "จำนวนชั่วโมงส่วนเกินในวันนี้:",
-        "month-balance": "จำนวนชั่วโมงส่วนเกินในเดือนนี้",
-        "month-balance-title": "จำนวนชั่วโมงส่วนเกินในเดือนนี้ ถ้าค่าเป็นบวกแสดงว่าทำเกินมา ถ้าค่าติดลบแสดงว่าทำขาดไป",
-        "overall-balance": "จำนวนชั่วโมงส่วนเกินทั้งหมด",
-        "overall-balance-title": "จำนวนชั่วโมงส่วนเกินทั้งหมด นับถึงวันปัจจุบันหรือสิ้นเดือนนี้",
-        "non-working-day": "ไม่ใช่วันทำงาน",
-        "waived-day": "วันหยุด",
-        "day-start": "เริ่มทำงาน",
-        "lunch-begin": "เริ่มพักเที่ยง",
-        "lunch-end": "สิ้นสุดพักเที่ยง",
-        "day-end": "สิ้นสุดการทำงาน",
-        "lunch-total": "รวมเวลาพักเที่ยง",
-        "day-total": "รวมเวลาทำงาน"
     },
     "$WorkdayWaiver": {
         "add-waiver": "เพิ่มวันหยุด",
