+++ locales/zh-TW/translation.json
@@ -44,7 +44,6 @@
         "clear-database": "清除資料庫",
         "confirm-clear-all-data": "確定清除所有資料？",
         "confirm-import-db": "確定導入資料庫？這會覆蓋所有現有資料。",
-        "confirm-migrate-to-flexible": "確定轉移固定日曆資料庫到自由日曆嗎？\n\n已有自由日曆資料庫將會被清除。",
         "copy": "複製",
         "could-not-be-loaded": "無法加載",
         "cut": "剪下",
@@ -59,14 +58,16 @@
         "export": "導出",
         "failed-entries": "紀錄時發生錯誤",
         "failed-migrating": "轉移時發生錯誤",
+        "fresh-start": "Fresh Start",
+        "fresh-start-confirm": "Are you sure you want a fresh start?",
         "help": "幫助",
         "import-database": "導入資料庫",
         "import-db-from-file": "從文件導入資料庫",
         "import-successful": "太好了！導入成功！",
         "import": "導入",
         "menu": "目錄",
+        "migrate": "Migrate",
         "migrate-calendar-to-flexible": "轉移固定日曆資料庫到自由",
-        "migrate-to-flexible-calendar": "轉移到自由日曆",
         "migration-successful": "太好了！轉移成功！",
         "no-thanks": "不用了，謝謝",
         "ok": "好的",
@@ -76,9 +77,10 @@
         "punched-time": "已打卡時間",
         "quit": "退出",
         "reload": "重新載入",
+        "report": "Report",
         "select-all": "選擇全部",
         "send-feedback": "發送意見反饋",
-        "should-migrate-to-flexible": "你的自由日曆為空。你想從已有固定日曆資料庫轉移到你的自由日曆嗎？",
+        "should-migrate-to-flexible": "One of the new features of this Time to Leave release is a flexible number of entries for the calendar.\n\nThis new view uses a new database format.\nTo use it, you can either migrate your existing database to the new format or make a fresh start.",
         "show-app": "顯示程式",
         "something-went-wrong": "出錯了 :(",
         "toggle-dev-tools": "切換開發者工具",
@@ -88,17 +90,11 @@
         "yes-please": "好的，謝謝"
     },
     "$FlexibleDayCalendar": {
-        "switch-view": "切換視角",
         "current-day": "回到當前日子",
         "previous-day": "上一日",
         "next-day": "下一日",
-        "time-to-leave": "Time to Leave",
         "leave-by": "你應該在此時間離開：",
-        "day-done-balance": "今日已全部完成，是日結餘：",
         "month-balance": "月份結餘",
-        "month-balance-title": "本月結餘會計算到今天為止，正結餘代表你今天（或今個月餘下的日子裡）有額外的小時數不用工作。",
-        "overall-balance": "總體結餘",
-        "overall-balance-title": "總體結餘直到月尾或當前日子",
         "not-a-working-day": "不是一個工作日",
         "waived-day": "已豁免日子",
         "day-total": "日子總計",
@@ -109,23 +105,17 @@
         "entry": "紀錄"
     },
     "$FlexibleMonthCalendar": {
-        "switch-view": "切換視角",
         "current-month": "回到當前月份",
         "previous-month": "上一個月",
         "next-month": "下一個月",
-        "time-to-leave": "Time to Leave",
         "day": "日",
         "total": "總計",
         "leave-by": "根據今天你的到達時間，你應在此時間離開",
-        "day-done-balance": "今日已全部完成，是日結餘：",
         "last-day-balance": "上一日已用結餘",
         "on": "在",
         "working-days": "工作日數",
         "working-days-title": "在這個月裡有多少個工作日",
         "month-balance": "月份結餘",
-        "month-balance-title": "本月結餘會計算到今天為止，正結餘代表你今天（或今個月餘下的日子裡）有額外的小時數不用工作。",
-        "overall-balance": "總體結餘",
-        "overall-balance-title": "總體結餘直到月尾或當前日子",
         "add-waiver-day": "為此日增加豁免",
         "waived-day": "已豁免日子",
         "remove-entry": "移除紀錄",
@@ -133,53 +123,18 @@
         "yes": "是",
         "no": "否"
     },
-    "$Calendar": {
-        "leave-by": "根據今天你的到達時間，你應在此時間離開",
-        "all-done": "今日已全部完成，是日結餘：",
-        "last-day-balance": "上一日已用結餘",
-        "on": "在",
-        "working-days": "工作日數",
-        "working-days-title": "在這個月裡有多少個工作日",
+    "$BaseCalendar": {
+        "day-done-balance": "今日已全部完成，是日結餘：",
         "month-balance": "月份結餘",
         "month-balance-title": "本月結餘會計算到今天為止，正結餘代表你今天（或今個月餘下的日子裡）有額外的小時數不用工作。",
         "overall-balance": "總體結餘",
         "overall-balance-title": "總體結餘直到月尾或當前日子",
         "switch-view": "切換視角",
-        "switch-view-title": "切換視角",
-        "current-month": "當前月份",
-        "current-month-title": "回到當前月份",
-        "previous-month": "上一個月",
-        "next-month": "下一個月",
-        "time-to-leave": "Time to Leave",
-        "day-start": "日子開始",
-        "lunch-begin": "午餐開始",
-        "lunch-end": "午餐結束",
-        "day-end": "日子結束",
-        "lunch-total": "午餐總計",
-        "day-total": "日子總計",
-        "day": "日",
-        "lunch-start": "午餐開始",
-        "punch-time": "打卡時間"
+        "time-to-leave": "Time to Leave"
     },
     "$Notification": {
         "time-to-leave": "你還好嗎？我想現在是時候離開了。",
         "punch-reminder": "別忘了打卡！"
-    },
-    "$FixedDayCalendar": {
-        "leave-by": "你應該在此時間離開：",
-        "all-done": "今日已全部完成，是日結餘：",
-        "month-balance": "月份結餘",
-        "month-balance-title": "本月結餘會計算到今天為止，正結餘代表你今天（或今個月餘下的日子裡）有額外的小時數不用工作。",
-        "overall-balance": "總體結餘",
-        "overall-balance-title": "總體結餘直到月尾或當前日子",
-        "non-working-day": "不是工作日",
-        "waived-day": "已豁免日子",
-        "day-start": "日子開始",
-        "lunch-begin": "午餐開始",
-        "lunch-end": "午餐結束",
-        "day-end": "日子結束",
-        "lunch-total": "午餐總計",
-        "day-total": "日子總計"
     },
     "$WorkdayWaiver": {
         "add-waiver": "加入豁免",
