+++ src/preferences.html
@@ -99,13 +99,6 @@
                 <label id='overall-balance-start-date'><input type="date" class="date-right" name="overall-balance-start-date"></label>
             </div>
             <div class="flex-box">
-                <p><i class="fas fa-align-center"></i><span data-i18n="$Preferences.numberOfEntries">Number of Entries</span></p>
-                <select class="selects-right" id="number-of-entries" name="number-of-entries">
-                    <option value="fixed" data-i18n="$Preferences.fixed" selected>Fixed</option>
-                    <option value="flexible" data-i18n="$Preferences.flexible">Flexible</option>
-                </select>
-            </div>
-            <div class="flex-box">
                 <p><i class="fas fa-eye"></i><span data-i18n="$Preferences.view">View</span></p>
                 <select class="selects-right" id="view" name="view">
                     <option value="month" data-i18n="$Preferences.month" selected>Month</option>
